Ask Optiq: close credential leaks, harden model-supplied input, split tools
The assistant sent its API key somewhere it should not have and trusted
numbers the model made up. Both are fixed here, and the tool layer is broken
up so the parts handling untrusted input can be reviewed on their own.

Credentials:
- Stop following redirects. cpp-httplib drops only a fixed set of headers on
  a cross-host redirect and the Azure subscription key is not one of them, so
  following one could hand the key to another host. A 3xx now reports why.
- Refuse to send the key over plain http unless the endpoint is loopback, so
  a local model server still works but a remote one cannot take it in clear.
- Elide query strings from logs and error text, in case a key was pasted in.
- Stop reading the pre-provider credential entry as a live fallback. Any
  endpoint without a key of its own inherited it, so a second endpoint at a
  different vendor would have been sent the first one's key. It is now a
  one-shot migration at load. Reset also clears the keys of endpoints it
  discards, instead of orphaning them in the credential store.

Model-supplied input:
- Reject non-finite time ranges. "end <= start" is false for NaN, so every
  range guard in the feature let NaN and infinity through.
- ShowPanel no longer reports success when there is no widget to change; it
  was moving global settings behind the user's back on compute traces.
- Clamp the SQL offset, bound JSON array arguments, escape LIKE wildcards so
  a name containing % or _ is matched literally, and reject malformed filters
  instead of quietly running a broader query than was asked for.
- goto could highlight nothing and still report success; it now navigates to
  an event that took and says so when none did. get_summary no longer formats
  an empty summary as though it were a measurement.
- Bookmark slots are range-checked on every entry point, not only on save.
- Tab lookup now needs a unique, long-enough match; a one-letter guess used
  to activate whichever tab happened to come first.

Structure:
- Split SQL fragment building into rocprofvis_ai_tool_query and the
  model-facing schema into rocprofvis_ai_tool_schema. The query module is the
  one place a bad argument can become bad SQL; the schema module touches no
  view state, which is what makes it safe on the HTTP worker thread.
- Replace the 1200-line StartAssistantTool if-chain with one function per
  tool behind a dispatch table. Bodies were moved verbatim.
- Fold four copies of a lowercase helper into to_lower_copy, and two copies
  of tab-name matching into TabContainer::FindTabByLabel.

Also fixes the main view losing ImGui state - column widths, tree expansion,
scroll position - every time the assistant panel was toggled, because it only
entered a child window when the panel was open.

Builds verified with ROCPROFVIS_ENABLE_AGENTIC_PROFILING both on and off.
