+++ CMakeLists.txt
@@ -241,6 +241,10 @@ list(APPEND VIEW_FILES
     src/view/src/agenticprofiling/rocprofvis_ai_assistant.h
     src/view/src/agenticprofiling/rocprofvis_ai_client.cpp
     src/view/src/agenticprofiling/rocprofvis_ai_client.h
+    src/view/src/agenticprofiling/rocprofvis_ai_tool_query.cpp
+    src/view/src/agenticprofiling/rocprofvis_ai_tool_query.h
+    src/view/src/agenticprofiling/rocprofvis_ai_tool_schema.cpp
+    src/view/src/agenticprofiling/rocprofvis_ai_tool_schema.h
     src/view/src/agenticprofiling/rocprofvis_ai_tools.cpp
     src/view/src/agenticprofiling/rocprofvis_ai_tools.h
 )
