+++ src/view/src/agenticprofiling/rocprofvis_ai_tools.cpp
@@ -5,6 +5,7 @@
 
 #include <algorithm>
 #include <cctype>
+#include <cmath>
 #include <cstdio>
 #include <cstdlib>
 #include <iomanip>
@@ -19,6 +20,8 @@
 #include "compute/rocprofvis_compute_selection.h"
 #include "model/compute/rocprofvis_compute_data_model.h"
 #include "rocprofvis_ai_actions.h"
+#include "rocprofvis_ai_tool_query.h"
+#include "rocprofvis_ai_tool_schema.h"
 #include "model/rocprofvis_analysis_model.h"
 #include "model/rocprofvis_common_defs.h"
 #include "model/rocprofvis_summary_model.h"
@@ -50,7 +53,6 @@ constexpr size_t   ASSISTANT_MAX_METRIC_NAMES = 40;
 constexpr size_t   ASSISTANT_MAX_METRIC_FETCH = 8;
 constexpr size_t   ASSISTANT_TOP_KERNEL_LIMIT = 10;
 constexpr size_t   ASSISTANT_MAX_RESULT_CHARS = 20000;
-constexpr size_t   ASSISTANT_MAX_FILTERS      = 8;
 constexpr size_t   ASSISTANT_MAX_CALL_STACK   = 40;
 constexpr size_t   ASSISTANT_MAX_FLOW_ROWS    = 20;
 constexpr size_t   ASSISTANT_MAX_EXT_ROWS     = 40;
@@ -64,47 +66,14 @@ constexpr size_t   ASSISTANT_MAX_NEXT_STEP_CHARS = 80;
 constexpr double   ASSISTANT_OVERVIEW_SCALE   = 100.0;
 constexpr uint64_t ASSISTANT_DURATION_COLUMN  = 2;
 
-// Columns the model is allowed to filter, group, and sort on. Filters reach the
-// database as a SQL fragment, so only names on this list are accepted and the
-// values beside them are quoted by QuoteSqlLiteral.
-const char* const ASSISTANT_QUERY_COLUMNS[] = {
-    "name",       "category",  "duration",   "start",      "end",
-    "id",         "__uuid",    "PID",        "TID",        "queue",
-    "stream",     "node",      "nodeId",     "size",       "address",
-    "SrcAddr",    "value",     "counter",    "arguments",  "GridSizeX",
-    "GridSizeY",  "GridSizeZ", "WGSizeX",    "WGSizeY",    "WGSizeZ",
-    "LDSSize",    "ScratchSize", "StaticLDSSize", "StaticScratchSize",
-    "AgentAbsoluteIndex", "AgentType", "AgentTypeIndex", "AgentName",
-    "SrcAgentAbsoluteIndex", "SrcAgentType", "SrcAgentTypeIndex",
-    "SrcAgentName", "__trackId", "__streamTrackId",
-};
+// Furthest a tool may page into a result set. limit is clamped too, but without
+// a ceiling here the model could ask the database to skip billions of rows.
+constexpr uint64_t ASSISTANT_MAX_OFFSET = 1000000;
 
-enum class QueryWildcard
-{
-    kNone,
-    kContains,
-    kPrefix
-};
-
-struct QueryOperator
-{
-    const char*   key;
-    const char*   sql;
-    QueryWildcard wildcard;
-};
-
-const QueryOperator ASSISTANT_QUERY_OPERATORS[] = {
-    { "=", "=", QueryWildcard::kNone },
-    { "==", "=", QueryWildcard::kNone },
-    { "!=", "!=", QueryWildcard::kNone },
-    { "<>", "!=", QueryWildcard::kNone },
-    { "<", "<", QueryWildcard::kNone },
-    { "<=", "<=", QueryWildcard::kNone },
-    { ">", ">", QueryWildcard::kNone },
-    { ">=", ">=", QueryWildcard::kNone },
-    { "contains", "LIKE", QueryWildcard::kContains },
-    { "starts_with", "LIKE", QueryWildcard::kPrefix },
-};
+// Longest JSON array a tool argument may be. The entries past a tool's own cap
+// are ignored anyway, so refusing outright is both cheaper and more honest than
+// silently reading the first few of a million.
+constexpr size_t ASSISTANT_MAX_ARRAY_ENTRIES = 256;
 
 struct TopEventsSpec
 {
@@ -113,38 +82,6 @@ struct TopEventsSpec
     TableType                          table_type;
     uint64_t                           request_id;
     rocprofvis_dm_event_operation_t    op;
-};
-
-struct AssistantToolLabel
-{
-    const char* name;
-    const char* status;
-};
-
-// Every tool the model can call, with the line the panel shows while it runs.
-// One list keeps the status text and the "unknown tool" reply in step with what
-// BuildAssistantToolsJson registers.
-const AssistantToolLabel ASSISTANT_TOOL_LABELS[] = {
-    { "trace_overview", "Reading the timeline overview..." },
-    { "get_summary", "Loading summary..." },
-    { "top_events", "Querying top events..." },
-    { "kernel_instances", "Loading kernel dispatches..." },
-    { "kernel_metrics", "Loading kernel metrics..." },
-    { "list_tracks", "Listing tracks..." },
-    { "search_events", "Searching the trace..." },
-    { "track_events", "Reading track events..." },
-    { "track_samples", "Reading counter samples..." },
-    { "event_details", "Loading event details..." },
-    { "track_statistics", "Computing track statistics..." },
-    { "goto", "Moving the view..." },
-    { "show_panel", "Opening a panel..." },
-    { "switch_tab", "Switching tabs..." },
-    { "flow_arrows", "Adjusting flow arrows..." },
-    { "annotate", "Leaving a note..." },
-    { "bookmark", "Working with bookmarks..." },
-    { "measure", "Measuring..." },
-    { "reset_view", "Resetting the view..." },
-    { "offer_next_steps", "Offering next steps..." },
 };
 
 const TopEventsSpec k_top_event_specs[] = {
@@ -176,17 +113,6 @@ Actions(const AssistantToolContext& context)
 {
     return OptiqActions(context.data_provider, context.timeline_selection,
                         context.compute_selection, context.trace_view);
-}
-
-// Lowercases a copy, for the case-insensitive matching the tools do on names.
-std::string
-ToLowerCopy(std::string value)
-{
-    for(char& c : value)
-    {
-        c = static_cast<char>(std::tolower(static_cast<unsigned char>(c)));
-    }
-    return value;
 }
 
 // Strips leading and trailing whitespace from a follow-up the model offered.
@@ -231,6 +157,29 @@ ClampRowLimit(int32_t requested)
         return ASSISTANT_DEFAULT_ROW_LIMIT;
     }
     return std::min(static_cast<size_t>(requested), ASSISTANT_MAX_ROW_LIMIT);
+}
+
+// Clamps how far a tool may page in, for the same reason as ClampRowLimit: the
+// number comes from the model and reaches the database unchanged otherwise.
+uint64_t
+ClampOffset(uint64_t requested)
+{
+    return std::min(requested, ASSISTANT_MAX_OFFSET);
+}
+
+// Rejects a model-supplied array that is longer than anything a tool can use.
+// Returns true when the array is fine to walk.
+bool
+CheckArrayLength(const std::vector<jt::Json>& entries, const char* argument_name,
+                 std::string& error_out)
+{
+    if(entries.size() <= ASSISTANT_MAX_ARRAY_ENTRIES)
+    {
+        return true;
+    }
+    error_out = std::string(argument_name) + " has too many entries. Use at most " +
+                std::to_string(ASSISTANT_MAX_ARRAY_ENTRIES) + ".";
+    return false;
 }
 
 // Formats a percentage to one decimal place, the way the UI shows it.
@@ -336,11 +285,19 @@ JsonU64(const jt::Json& json, const std::string& key, uint64_t default_value)
     return default_value;
 }
 
+// The row offset a paging tool was asked for, bounded. Every table read goes
+// through here so none of them can forget to clamp.
+uint64_t
+OffsetFromArgs(const jt::Json& args)
+{
+    return ClampOffset(JsonU64(args, "offset", 0));
+}
+
 // Maps a category name, plus the synonyms the model likes, to its table spec.
 const TopEventsSpec*
 FindTopEventsSpec(const std::string& category)
 {
-    const std::string lowered = ToLowerCopy(category);
+    const std::string lowered = to_lower_copy(category);
     std::string       key     = lowered;
     if(key == "kernel" || key == "kernels" || key == "dispatch_events")
     {
@@ -487,229 +444,6 @@ KeepTracksOfType(const AssistantToolContext& context, const std::vector<uint64_t
     return kept;
 }
 
-// Reports whether a column name is one a tool argument is allowed to name.
-bool
-IsAllowedQueryColumn(const std::string& name)
-{
-    for(const char* column : ASSISTANT_QUERY_COLUMNS)
-    {
-        if(name == column)
-        {
-            return true;
-        }
-    }
-    return false;
-}
-
-// Lists the allowed columns, so a rejection tells the model what to use.
-std::string
-AllowedQueryColumnList()
-{
-    std::ostringstream out;
-    bool               first = true;
-    for(const char* column : ASSISTANT_QUERY_COLUMNS)
-    {
-        if(!first)
-        {
-            out << ", ";
-        }
-        first = false;
-        out << column;
-    }
-    return out.str();
-}
-
-// Looks up a filter operator, so only known spellings reach the query.
-const QueryOperator*
-FindQueryOperator(const std::string& key)
-{
-    for(const QueryOperator& op : ASSISTANT_QUERY_OPERATORS)
-    {
-        if(key == op.key)
-        {
-            return &op;
-        }
-    }
-    return nullptr;
-}
-
-// Renders a string as a SQL literal: single quotes doubled, control characters
-// dropped, so a value cannot terminate the literal and inject syntax.
-std::string
-QuoteSqlLiteral(const std::string& value)
-{
-    std::string quoted = "'";
-    for(char c : value)
-    {
-        if(static_cast<unsigned char>(c) < 0x20)
-        {
-            continue;
-        }
-        if(c == '\'')
-        {
-            quoted += '\'';
-        }
-        quoted += c;
-    }
-    quoted += '\'';
-    return quoted;
-}
-
-// Prints a double at full precision so a filter value survives the trip.
-std::string
-FormatNumberLiteral(double value)
-{
-    std::ostringstream out;
-    out << std::setprecision(17) << value;
-    return out.str();
-}
-
-// Turns a validated filter list into a boolean SQL fragment. The query builder
-// appends this after "AND", so it must not contain the WHERE keyword.
-std::string
-BuildWhereClause(const jt::Json& args, std::string& error_out)
-{
-    jt::Json& mutable_args = const_cast<jt::Json&>(args);
-    if(!mutable_args.contains("filters") || !mutable_args["filters"].isArray())
-    {
-        return std::string();
-    }
-
-    std::vector<jt::Json>& entries = mutable_args["filters"].getArray();
-    if(entries.size() > ASSISTANT_MAX_FILTERS)
-    {
-        error_out = "Too many filters. Use at most " +
-                    std::to_string(ASSISTANT_MAX_FILTERS) + ".";
-        return std::string();
-    }
-
-    std::ostringstream out;
-    size_t             used = 0;
-    for(jt::Json& entry : entries)
-    {
-        if(!entry.isObject())
-        {
-            continue;
-        }
-
-        const std::string column = JsonUtils::GetString(entry, "column", "");
-        if(!IsAllowedQueryColumn(column))
-        {
-            error_out = "Unknown filter column \"" + column +
-                        "\". Allowed columns: " + AllowedQueryColumnList();
-            return std::string();
-        }
-
-        const QueryOperator* op =
-            FindQueryOperator(ToLowerCopy(JsonUtils::GetString(entry, "op", "=")));
-        if(op == nullptr)
-        {
-            error_out = "Unknown filter op. Use =, !=, <, <=, >, >=, contains, "
-                        "or starts_with.";
-            return std::string();
-        }
-
-        if(!entry.contains("value"))
-        {
-            error_out = "Filter on \"" + column + "\" is missing value.";
-            return std::string();
-        }
-
-        jt::Json&   value = entry["value"];
-        std::string literal;
-        if(value.isString())
-        {
-            std::string text = value.getString();
-            if(op->wildcard == QueryWildcard::kContains)
-            {
-                text = "%" + text + "%";
-            }
-            else if(op->wildcard == QueryWildcard::kPrefix)
-            {
-                text += "%";
-            }
-            literal = QuoteSqlLiteral(text);
-        }
-        else if(value.isLong())
-        {
-            literal = std::to_string(value.getLong());
-        }
-        else if(value.isDouble())
-        {
-            literal = FormatNumberLiteral(value.getDouble());
-        }
-        else
-        {
-            error_out = "Filter on \"" + column + "\" needs a string or number value.";
-            return std::string();
-        }
-
-        if(used > 0)
-        {
-            out << " AND ";
-        }
-        out << column << " " << op->sql << " " << literal;
-        ++used;
-    }
-    return out.str();
-}
-
-// group_by goes into the query as a bare column name, so it gets the same
-// whitelist treatment as filter columns.
-std::string
-GroupByFromArgs(const jt::Json& args, std::string& error_out)
-{
-    const std::string group = JsonUtils::GetString(args, "group_by", "");
-    if(group.empty() || !error_out.empty())
-    {
-        return std::string();
-    }
-    if(!IsAllowedQueryColumn(group))
-    {
-        error_out = "Unknown group_by column \"" + group +
-                    "\". Allowed columns: " + AllowedQueryColumnList();
-        return std::string();
-    }
-    return group;
-}
-
-// Turns a sort_by column name into its index in the table's own header.
-uint64_t
-ResolveSortColumn(const TablesModel& tables, TableType type, const std::string& name,
-                  uint64_t fallback)
-{
-    if(name.empty())
-    {
-        return fallback;
-    }
-    const std::vector<std::string>& header  = tables.GetTableHeader(type);
-    const std::string               lowered = ToLowerCopy(name);
-    for(size_t i = 0; i < header.size(); ++i)
-    {
-        if(ToLowerCopy(header[i]) == lowered)
-        {
-            return static_cast<uint64_t>(i);
-        }
-    }
-    return fallback;
-}
-
-// Reads asc/desc, keeping each tool's own default when the model omits it.
-rocprofvis_controller_sort_order_t
-SortOrderFromArgs(const jt::Json& args, rocprofvis_controller_sort_order_t fallback)
-{
-    const std::string order = ToLowerCopy(JsonUtils::GetString(args, "sort_order", ""));
-    if(order == "asc" || order == "ascending")
-    {
-        return kRPVControllerSortOrderAscending;
-    }
-    if(order == "desc" || order == "descending")
-    {
-        return kRPVControllerSortOrderDescending;
-    }
-    return fallback;
-}
-
 // Explicit start_ns/end_ns win, else the user's selection, else the whole
 // trace. Search passes prefer_selection false to match its "Whole Trace" default.
 void
@@ -729,11 +463,28 @@ TimeRangeFromArgs(const AssistantToolContext& context, const jt::Json& args,
 
     const double requested_start = JsonUtils::GetDouble(args, "start_ns", -1.0);
     const double requested_end   = JsonUtils::GetDouble(args, "end_ns", -1.0);
-    if(requested_start >= 0.0 && requested_end > requested_start)
+    // std::isfinite rejects the infinities; NaN fails the comparisons below on
+    // its own, since every comparison against NaN is false.
+    if(!std::isfinite(requested_start) || !std::isfinite(requested_end) ||
+       requested_start < 0.0 || requested_end <= requested_start)
     {
-        start_ns = requested_start;
-        end_ns   = requested_end;
+        return;
     }
+
+    // Keep the window inside the trace. A range the model invented out past the
+    // end would otherwise come back empty and read as "nothing happened here"
+    // rather than "you asked about time that was never recorded".
+    const TimelineModel& timeline = context.data_provider->DataModel().GetTimeline();
+    const double         trace_start = timeline.GetStartTime();
+    const double         trace_end   = timeline.GetEndTime();
+    const double clamped_start = std::max(requested_start, trace_start);
+    const double clamped_end   = std::min(requested_end, trace_end);
+    if(clamped_end <= clamped_start)
+    {
+        return;
+    }
+    start_ns = clamped_start;
+    end_ns   = clamped_end;
 }
 
 // Track ids named by the model, filtered to tracks that exist and carry the
@@ -749,10 +500,15 @@ TracksFromArgs(const AssistantToolContext& context, const jt::Json& args,
         return require_op ? TracksForOperation(context, op) : std::vector<uint64_t>();
     }
 
-    const TimelineModel&   timeline = context.data_provider->DataModel().GetTimeline();
-    std::vector<jt::Json>& entries  = mutable_args["track_ids"].getArray();
-    std::vector<uint64_t>  tracks;
-    tracks.reserve(entries.size());
+    std::vector<jt::Json>& entries = mutable_args["track_ids"].getArray();
+    if(!CheckArrayLength(entries, "track_ids", error_out))
+    {
+        return std::vector<uint64_t>();
+    }
+
+    const TimelineModel&  timeline = context.data_provider->DataModel().GetTimeline();
+    std::vector<uint64_t> tracks;
+    tracks.reserve(std::min(entries.size(), ASSISTANT_MAX_TRACKS));
     for(jt::Json& entry : entries)
     {
         uint64_t track_id = 0;
@@ -766,7 +522,8 @@ TracksFromArgs(const AssistantToolContext& context, const jt::Json& args,
         }
         else
         {
-            continue;
+            error_out = "Every track_ids entry must be a number.";
+            return std::vector<uint64_t>();
         }
 
         const TrackInfo* track = timeline.GetTrack(track_id);
@@ -1323,14 +1080,14 @@ FindComputeKernel(const WorkloadInfo& workload, const std::string& name, uint32_
             return kernel;
         }
     }
-    const std::string lowered = ToLowerCopy(name);
+    const std::string lowered = to_lower_copy(name);
     for(const KernelInfo* kernel : workload.ordered_kernels)
     {
         if(kernel == nullptr)
         {
             continue;
         }
-        if(ToLowerCopy(kernel->name).find(lowered) != std::string::npos)
+        if(to_lower_copy(kernel->name).find(lowered) != std::string::npos)
         {
             return kernel;
         }
@@ -1384,10 +1141,17 @@ OfferNextStepsResult(const jt::Json& args)
                           "Missing steps");
     }
 
+    std::vector<jt::Json>& entries = mutable_args["steps"].getArray();
+    std::string            length_error;
+    if(!CheckArrayLength(entries, "steps", length_error))
+    {
+        return DoneResult(length_error, "Too many steps");
+    }
+
     AssistantToolStartResult result =
         DoneResult("Offered next steps under the chat.", "Offered next steps");
     result.set_next_steps = true;
-    for(jt::Json& entry : mutable_args["steps"].getArray())
+    for(jt::Json& entry : entries)
     {
         if(!entry.isString())
         {
@@ -1450,390 +1214,8 @@ PendingResult(uint64_t request_id, const AssistantFetchState& fetch,
                          started_fetch);
 }
 
-// Builds a JSON string array, for the enum of a tool parameter.
-jt::Json
-MakeStringEnum(const std::vector<const char*>& values)
-{
-    jt::Json array;
-    for(size_t i = 0; i < values.size(); ++i)
-    {
-        array[i] = values[i];
-    }
-    return array;
-}
-
-// Registers one tool in the schema array the model receives.
-void
-AddTool(jt::Json& tools, size_t index, const char* name, const char* description,
-        jt::Json parameters)
-{
-    tools[index]["type"]                    = "function";
-    tools[index]["function"]["name"]        = name;
-    tools[index]["function"]["description"] = description;
-    tools[index]["function"]["parameters"]  = parameters;
-}
-
-// Starts an empty object schema, which every tool's parameters build on.
-jt::Json
-ObjectParams()
-{
-    jt::Json params;
-    params["type"] = "object";
-    params["properties"].setObject();
-    return params;
-}
-
-// Adds one typed, described property to a parameter schema.
-void
-AddParam(jt::Json& params, const char* name, const char* type, const char* description)
-{
-    params["properties"][name]["type"]        = type;
-    params["properties"][name]["description"] = description;
-}
-
-// The window and tracks a table tool reads from.
-void
-AddScopeParams(jt::Json& params)
-{
-    AddParam(params, "start_ns", "number",
-             "Window start in nanoseconds. Defaults to the user's selected range, "
-             "or the whole trace.");
-    AddParam(params, "end_ns", "number", "Window end in nanoseconds.");
-    AddParam(params, "track_ids", "array",
-             "Track ids from list_tracks. Defaults to every track that carries the "
-             "requested events.");
-    params["properties"]["track_ids"]["items"]["type"] = "integer";
-}
-
-// The structured stand-in for a SQL query: filters, sort, paging, grouping.
-void
-AddQueryParams(jt::Json& params)
-{
-    jt::Json filter_item = ObjectParams();
-    AddParam(filter_item, "column", "string",
-             "Column name. A bad name comes back with the allowed list.");
-    AddParam(filter_item, "op", "string",
-             "=, !=, <, <=, >, >=, contains, or starts_with.");
-    filter_item["properties"]["value"]["description"] =
-        "String or number to compare against.";
-    filter_item["required"][0] = "column";
-    filter_item["required"][1] = "value";
-
-    AddParam(params, "filters", "array",
-             "Conditions combined with AND, e.g. "
-             "[{\"column\":\"duration\",\"op\":\">=\",\"value\":5000}].");
-    params["properties"]["filters"]["items"] = filter_item;
-
-    AddParam(params, "sort_by", "string",
-             "Column name to sort by. Call once without it to see the columns.");
-    AddParam(params, "sort_order", "string", "asc or desc.");
-    params["properties"]["sort_order"]["enum"] = MakeStringEnum({ "asc", "desc" });
-    AddParam(params, "limit", "integer", "Rows to return (default 20, max 200).");
-    AddParam(params, "offset", "integer",
-             "Rows to skip. Use with limit to page through a big result.");
-    AddParam(params, "group_by", "string",
-             "Optional column to aggregate rows by, e.g. name.");
-}
-
-// Every registered tool name, for telling the model when it invents one.
-std::string
-AssistantToolNameList()
-{
-    std::ostringstream out;
-    bool               first = true;
-    for(const AssistantToolLabel& label : ASSISTANT_TOOL_LABELS)
-    {
-        if(!first)
-        {
-            out << ", ";
-        }
-        first = false;
-        out << label.name;
-    }
-    return out.str();
-}
-
 }  // namespace
 
-// Builds the tool schema sent with every request. The descriptions here are
-// the only instructions the model gets about what each tool is for.
-jt::Json
-BuildAssistantToolsJson()
-{
-    jt::Json tools;
-
-    jt::Json summary_params = ObjectParams();
-    AddTool(tools, 0, "get_summary",
-            "Load named GPU summary stats and the top kernels by execution time. "
-            "Call this first when the briefing is empty or kernel_exec_time_total_ns is 0.",
-            summary_params);
-
-    jt::Json top_params = ObjectParams();
-    AddParam(top_params, "category", "string", "Event family to rank by duration.");
-    top_params["properties"]["category"]["enum"] =
-        MakeStringEnum({ "dispatch", "memory_copy", "memory_alloc", "instrumented",
-                         "sampled" });
-    AddScopeParams(top_params);
-    AddQueryParams(top_params);
-    top_params["required"][0] = "category";
-    AddTool(tools, 1, "top_events",
-            "Rank the hottest events of one category, optionally narrowed to named "
-            "tracks, an explicit time window, and column filters. Uses the same "
-            "analysis tables as View > Top Events. Does not run raw SQL.",
-            top_params);
-
-    jt::Json inst_params = ObjectParams();
-    AddParam(inst_params, "kernel_name", "string",
-             "Exact or unique kernel name from get_summary / top_events.");
-    AddScopeParams(inst_params);
-    AddQueryParams(inst_params);
-    inst_params["required"][0] = "kernel_name";
-    AddTool(tools, 2, "kernel_instances",
-            "List individual GPU dispatches for one kernel name, with timestamps "
-            "and track ids that goto can use.",
-            inst_params);
-
-    jt::Json metrics_params = ObjectParams();
-    AddParam(metrics_params, "kernel_name", "string",
-             "Kernel name (compute traces) or summary kernel name (system traces).");
-    AddParam(metrics_params, "kernel_id", "integer",
-             "Kernel id from get_summary, instead of the name, on a compute trace.");
-    AddParam(metrics_params, "metric_name", "string",
-             "Optional substring of a hardware metric to fetch on a compute trace.");
-    AddTool(tools, 3, "kernel_metrics",
-            "Return duration/call stats for a kernel. On compute traces, can also "
-            "fetch named hardware metrics. On system traces, returns summary stats.",
-            metrics_params);
-
-    jt::Json tracks_params = ObjectParams();
-    AddTool(tools, 4, "list_tracks",
-            "List timeline tracks with ids, names, and event families. Use before goto "
-            "if you need a track_id.",
-            tracks_params);
-
-    jt::Json goto_params = ObjectParams();
-    AddParam(goto_params, "start_ns", "number", "Range start in nanoseconds.");
-    AddParam(goto_params, "end_ns", "number", "Range end in nanoseconds.");
-    AddParam(goto_params, "track_id", "integer",
-             "__trackId of the event. Pass it whenever you pass event_uuid; without "
-             "it the flow arrows cannot be drawn.");
-    AddParam(goto_params, "event_uuid", "integer",
-             "__uuid of the event to select. Selecting it expands its flow arrows "
-             "and call stack, the same as clicking it.");
-    AddParam(goto_params, "kernel_name", "string",
-             "On compute traces, select this kernel in the UI.");
-    AddParam(goto_params, "kernel_id", "integer",
-             "On compute traces, select this kernel id instead of the name.");
-
-    jt::Json goto_event_item = ObjectParams();
-    AddParam(goto_event_item, "track_id", "integer", "Track the event lives on.");
-    AddParam(goto_event_item, "event_uuid", "integer", "__uuid of the event.");
-    goto_event_item["required"][0] = "event_uuid";
-    AddParam(goto_params, "events", "array",
-             "Events to light up on the timeline, from the __trackId and __uuid "
-             "columns. They stay highlighted while the user reads, and the first "
-             "one gets selected so its flow arrows and call stack expand.");
-    goto_params["properties"]["events"]["items"] = goto_event_item;
-    AddParam(goto_params, "zoom", "boolean",
-             "Also zoom the timeline to the range, not just select it. Use this "
-             "when the range is small enough that the user would have to zoom in "
-             "by hand.");
-
-    AddTool(tools, 5, "goto",
-            "Point the UI at what you are talking about: move the timeline to a "
-            "range, highlight the specific events behind your claim, and expand "
-            "the first one's connecting arrows. Call this on your own whenever you "
-            "name a window or an outlier, so the user sees exactly what you saw. "
-            "A range on its own only scrolls the view - pass event_uuid and "
-            "track_id, or nothing gets highlighted and no arrows appear. Do not "
-            "use it to pin notes or change other UI.",
-            goto_params);
-
-    jt::Json track_events_params = ObjectParams();
-    AddScopeParams(track_events_params);
-    AddQueryParams(track_events_params);
-    AddTool(tools, 6, "track_events",
-            "Read raw events off specific timeline tracks. This is the closest thing "
-            "to querying the database: pick tracks with track_ids, a window with "
-            "start_ns/end_ns, narrow rows with filters, then sort and page. Returns "
-            "the same rows as the Event Table tab.",
-            track_events_params);
-
-    jt::Json track_samples_params = ObjectParams();
-    AddScopeParams(track_samples_params);
-    AddQueryParams(track_samples_params);
-    AddTool(tools, 7, "track_samples",
-            "Read counter/PMC samples off specific counter tracks over a window. "
-            "Use track_ids from list_tracks where type=Counter. Same filtering, "
-            "sorting, and paging as track_events.",
-            track_samples_params);
-
-    jt::Json event_details_params = ObjectParams();
-    AddParam(event_details_params, "event_uuid", "integer",
-             "Event uuid from the __uuid column of track_events, top_events, or "
-             "kernel_instances.");
-    AddParam(event_details_params, "track_id", "integer",
-             "Optional owning track id, which fills in the event name and duration.");
-    event_details_params["required"][0] = "event_uuid";
-    AddTool(tools, 8, "event_details",
-            "Drill into one event: its arguments, extended data, flow links to and "
-            "from other events, and its call stack. Use after track_events or "
-            "top_events gives you a __uuid.",
-            event_details_params);
-
-    jt::Json track_stats_params = ObjectParams();
-    AddParam(track_stats_params, "track_id", "integer",
-             "Queue track (returns utilization) or counter track (returns "
-             "min/max/mean/stddev). From list_tracks.");
-    AddParam(track_stats_params, "start_ns", "number",
-             "Window start in nanoseconds. Defaults to the selection or whole trace.");
-    AddParam(track_stats_params, "end_ns", "number", "Window end in nanoseconds.");
-    track_stats_params["required"][0] = "track_id";
-    AddTool(tools, 9, "track_statistics",
-            "Compute statistics for one track over a window: queue busy percentage "
-            "for a queue track, or min/max/mean/standard deviation for a counter "
-            "track. Answers \"how busy was this GPU\" and \"how hot did it get\".",
-            track_stats_params);
-
-    jt::Json search_params = ObjectParams();
-    AddParam(search_params, "terms", "array",
-             "Names to search for, e.g. [\"gemm\", \"memcpy\"].");
-    search_params["properties"]["terms"]["items"]["type"] = "string";
-    AddParam(search_params, "match", "string",
-             "contains (default, substring) or equals (exact name).");
-    search_params["properties"]["match"]["enum"] =
-        MakeStringEnum({ "contains", "equals" });
-    AddParam(search_params, "combine", "string",
-             "all (default, an event must match every term) or any.");
-    search_params["properties"]["combine"]["enum"] = MakeStringEnum({ "all", "any" });
-    AddParam(search_params, "include_category", "boolean",
-             "Also match against event category names. Default false.");
-    AddParam(search_params, "categories", "array",
-             "Limit to these event families. Defaults to all of them.");
-    search_params["properties"]["categories"]["items"]["type"] = "string";
-    search_params["properties"]["categories"]["items"]["enum"] =
-        MakeStringEnum({ "dispatch", "memory_copy", "memory_alloc", "instrumented",
-                         "sampled" });
-    AddParam(search_params, "start_ns", "number",
-             "Window start in nanoseconds. Defaults to the whole trace.");
-    AddParam(search_params, "end_ns", "number", "Window end in nanoseconds.");
-    AddQueryParams(search_params);
-    search_params["required"][0] = "terms";
-    AddTool(tools, 10, "search_events",
-            "Search the whole trace by event name, the same as the search box in "
-            "the toolbar. Use this when you know part of a name but not which track "
-            "or when it ran. Returns matching events with track ids, timestamps, and "
-            "__uuid values you can pass to goto or event_details.",
-            search_params);
-
-    jt::Json overview_params = ObjectParams();
-    AddParam(overview_params, "track_id", "integer",
-             "Optional: profile one track instead of the whole trace.");
-    AddParam(overview_params, "bins", "integer",
-             "Time buckets to report (default 32).");
-    AddTool(tools, 11, "trace_overview",
-            "Preliminary analysis. Reads the timeline histogram and minimap that "
-            "Optiq already built: where activity is concentrated in time, the "
-            "busiest and quietest windows as real nanosecond ranges, total event "
-            "count, and which tracks own the most events. Costs no database query. "
-            "Call this FIRST so later tools can target a window and a track instead "
-            "of scanning the whole trace.",
-            overview_params);
-
-    jt::Json panel_params = ObjectParams();
-    AddParam(panel_params, "panel", "string",
-             ("Which panel: " + OptiqActions::PanelNameList()).c_str());
-    AddParam(panel_params, "visible", "boolean",
-             "true to open, false to close. Defaults to true.");
-    panel_params["required"][0] = "panel";
-    AddTool(tools, 12, "show_panel",
-            "Open or close one of Optiq's panels, the same as the View menu. Only "
-            "call this when the user asked you to show or hide that panel.",
-            panel_params);
-
-    jt::Json tab_params = ObjectParams();
-    AddParam(tab_params, "name", "string",
-             "Tab to switch to. Part of the name is enough. Omit to list every "
-             "tab that is available.");
-    AddTool(tools, 13, "switch_tab",
-            "Switch tabs among open traces or the details panel. Only call this "
-            "when the user asked you to change tabs. Call with no name to list "
-            "what is available.",
-            tab_params);
-
-    jt::Json flow_params = ObjectParams();
-    AddParam(flow_params, "visible", "boolean",
-             "Show or hide the flow arrows. Defaults to true.");
-    AddParam(flow_params, "style", "string",
-             "fan draws every link from the selected event; chain follows the "
-             "sequence end to end.");
-    flow_params["properties"]["style"]["enum"] = MakeStringEnum({ "fan", "chain" });
-    AddTool(tools, 14, "flow_arrows",
-            "Show, hide, or restyle the arrows that connect an event to the events "
-            "it launched or waited on. Selecting an event with goto expands that "
-            "event's arrows, but only while arrows are switched on, so pair this "
-            "with goto using visible=true whenever you select an event - the user "
-            "may have turned them off. Wait to be asked only before restyling them "
-            "or switching them off.",
-            flow_params);
-
-    jt::Json note_params = ObjectParams();
-    AddParam(note_params, "time_ns", "number",
-             "Where on the timeline to pin the note.");
-    AddParam(note_params, "title", "string", "Short heading, a few words.");
-    AddParam(note_params, "text", "string",
-             "What you found here and why it matters.");
-    AddParam(note_params, "track_id", "integer",
-             "Optional track to attach the note to.");
-    note_params["required"][0] = "time_ns";
-    note_params["required"][1] = "title";
-    note_params["required"][2] = "text";
-    AddTool(tools, 15, "annotate",
-            "Pin a sticky note on the timeline. Notes are saved with the project. "
-            "Only call this when the user asked you to leave a note.",
-            note_params);
-
-    jt::Json bookmark_params = ObjectParams();
-    AddParam(bookmark_params, "action", "string",
-             "save the current view, go to a saved one, remove one, or list them.");
-    bookmark_params["properties"]["action"]["enum"] =
-        MakeStringEnum({ "save", "goto", "remove", "list" });
-    AddParam(bookmark_params, "slot", "integer", "Bookmark slot, 0 to 9.");
-    AddTool(tools, 16, "bookmark",
-            "Save, jump to, remove, or list numbered zoom bookmarks. Only call "
-            "this when the user asked you to work with bookmarks.",
-            bookmark_params);
-
-    jt::Json measure_params = ObjectParams();
-    AddParam(measure_params, "start_ns", "number", "Span start in nanoseconds.");
-    AddParam(measure_params, "end_ns", "number", "Span end in nanoseconds.");
-    AddParam(measure_params, "clear", "boolean",
-             "Set true to clear the measurement instead of taking one.");
-    AddTool(tools, 17, "measure",
-            "Drop the two measurement pins on a span so the toolbar shows its "
-            "duration. Only call this when the user asked you to measure a span.",
-            measure_params);
-
-    jt::Json reset_params = ObjectParams();
-    AddTool(tools, 18, "reset_view",
-            "Zoom the timeline back out to the whole trace. Only call this when "
-            "the user asked you to reset the view.",
-            reset_params);
-
-    jt::Json next_params = ObjectParams();
-    AddParam(next_params, "steps", "array",
-             "Two or three short follow-ups the user can click, most useful first. "
-             "Each is a complete thing they would type, under 80 characters.");
-    next_params["properties"]["steps"]["items"]["type"] = "string";
-    next_params["required"][0] = "steps";
-    AddTool(tools, 19, "offer_next_steps",
-            "Puts stacked buttons under the chat for what to look at next. Call it "
-            "as the last tool of an investigation, then write your answer in the "
-            "response after it. Do not list those same options in the prose.",
-            next_params);
-
-    return tools;
-}
 
 // Builds the short trace description the model sees before it asks anything.
 std::string
@@ -2027,19 +1409,1217 @@ GetAssistantActivityRows(const AssistantToolContext& context, size_t bin_count,
     return rows;
 }
 
-// The line the panel shows under the transcript while a tool runs.
-std::string
-AssistantToolStatusLabel(const std::string& tool_name)
+
+namespace
 {
-    for(const AssistantToolLabel& label : ASSISTANT_TOOL_LABELS)
+
+// One tool body. Each takes the context for the trace in front and the
+// already-parsed arguments; the name it was called by is passed too, but only
+// the shared track_events / track_samples body has any use for it.
+using AssistantToolHandler = AssistantToolStartResult (*)(
+    const AssistantToolContext& context, const jt::Json& args,
+    const std::string& tool_name);
+
+// Implements the offer_next_steps tool. The only one that runs without a trace,
+// since it just puts buttons under the chat.
+AssistantToolStartResult
+ToolOfferNextSteps(const AssistantToolContext&, const jt::Json& args,
+                   const std::string&)
+{
+    return OfferNextStepsResult(args);
+}
+
+// Implements the show_panel tool.
+AssistantToolStartResult
+ToolShowPanel(const AssistantToolContext& context, const jt::Json& args,
+              const std::string&)
+{
+    const std::string panel_name = JsonUtils::GetString(args, "panel", "");
+    const OptiqPanel  panel      = OptiqActions::PanelFromName(panel_name);
+    if(panel == OptiqPanel::kUnknown)
     {
-        if(tool_name == label.name)
+        return DoneResult("Unknown panel \"" + panel_name + "\". Use one of: " +
+                              OptiqActions::PanelNameList() + ".",
+                          "Unknown panel");
+    }
+
+    const bool visible = JsonUtils::GetBool(args, "visible", true);
+    if(!Actions(context).ShowPanel(panel, visible))
+    {
+        return DoneResult(std::string("Could not change the ") +
+                              OptiqActions::PanelName(panel) + " panel here.",
+                          "Panel unavailable");
+    }
+    return DoneResult(std::string(visible ? "Opened the " : "Closed the ") +
+                          OptiqActions::PanelName(panel) + " panel.",
+                      visible ? "Opened a panel" : "Closed a panel");
+}
+
+// Implements the reset_view tool.
+AssistantToolStartResult
+ToolResetView(const AssistantToolContext& context, const jt::Json& args,
+              const std::string&)
+{
+    if(!Actions(context).ResetView())
+    {
+        return DoneResult("Reset view is only available on a system trace.",
+                          "Not available");
+    }
+    return DoneResult("Zoomed back out to the whole trace.", "Reset the view");
+}
+
+// Implements the measure tool.
+AssistantToolStartResult
+ToolMeasure(const AssistantToolContext& context, const jt::Json& args,
+            const std::string&)
+{
+    OptiqActions actions = Actions(context);
+    if(JsonUtils::GetBool(args, "clear", false))
+    {
+        actions.ClearMeasurement();
+        return DoneResult("Cleared the measurement.", "Cleared measurement");
+    }
+
+    const double start_ns = JsonUtils::GetDouble(args, "start_ns", -1.0);
+    const double end_ns   = JsonUtils::GetDouble(args, "end_ns", -1.0);
+    if(start_ns < 0.0 || end_ns <= start_ns)
+    {
+        return DoneResult("measure needs start_ns and end_ns with end > start.",
+                          "Bad range");
+    }
+    if(!actions.MeasureRange(start_ns, end_ns))
+    {
+        return DoneResult("Could not measure on this trace.", "Not available");
+    }
+
+    SettingsManager& settings    = SettingsManager::GetInstance();
+    const TimeFormat time_format = settings.GetUserSettings().unit_settings.time_format;
+    return DoneResult(
+        "Measured " +
+            nanosecond_to_formatted_str(end_ns - start_ns, time_format, true) +
+            " on the timeline.",
+        "Measured a span");
+}
+
+// Implements the bookmark tool.
+AssistantToolStartResult
+ToolBookmark(const AssistantToolContext& context, const jt::Json& args,
+             const std::string&)
+{
+    OptiqActions      actions = Actions(context);
+    const std::string action  = to_lower_copy(JsonUtils::GetString(args, "action", "list"));
+    const std::vector<int> slots = actions.ListBookmarks();
+
+    std::ostringstream saved;
+    saved << "saved_bookmarks:";
+    if(slots.empty())
+    {
+        saved << " none";
+    }
+    for(int slot : slots)
+    {
+        saved << " " << slot;
+    }
+    saved << "\n";
+
+    if(action == "list")
+    {
+        return DoneResult(saved.str(), "Listed bookmarks");
+    }
+
+    const int slot = static_cast<int>(JsonUtils::GetInt(args, "slot", -1));
+    if(slot < 0)
+    {
+        return DoneResult("bookmark needs a slot from 0 to 9.\n" + saved.str(),
+                          "Missing slot");
+    }
+    if(action == "save")
+    {
+        if(!actions.SaveBookmark(slot))
         {
-            return label.status;
+            return DoneResult("Could not save that bookmark.", "Bookmark failed");
+        }
+        return DoneResult("Saved the current view to bookmark " +
+                              std::to_string(slot) + ".",
+                          "Saved a bookmark");
+    }
+    if(action == "goto")
+    {
+        if(!actions.GotoBookmark(slot))
+        {
+            return DoneResult("Bookmark " + std::to_string(slot) +
+                                  " is empty.\n" + saved.str(),
+                              "No such bookmark");
+        }
+        return DoneResult("Jumped to bookmark " + std::to_string(slot) + ".",
+                          "Used a bookmark");
+    }
+    if(action == "remove")
+    {
+        if(!actions.RemoveBookmark(slot))
+        {
+            return DoneResult("Bookmark " + std::to_string(slot) + " is empty.",
+                              "No such bookmark");
+        }
+        return DoneResult("Removed bookmark " + std::to_string(slot) + ".",
+                          "Removed a bookmark");
+    }
+    return DoneResult("Unknown action. Use save, goto, remove, or list.",
+                      "Bad action");
+}
+
+// Implements the annotate tool.
+AssistantToolStartResult
+ToolAnnotate(const AssistantToolContext& context, const jt::Json& args,
+             const std::string&)
+{
+    if(context.is_compute)
+    {
+        return DoneResult("Notes are a timeline feature, so system traces only.",
+                          "Wrong trace kind");
+    }
+    const double      time_ns = JsonUtils::GetDouble(args, "time_ns", -1.0);
+    const std::string title   = JsonUtils::GetString(args, "title", "");
+    const std::string text    = JsonUtils::GetString(args, "text", "");
+    if(time_ns < 0.0 || title.empty() || text.empty())
+    {
+        return DoneResult("annotate needs time_ns, title, and text.",
+                          "Missing arguments");
+    }
+
+    double view_start = 0.0;
+    double view_end   = 0.0;
+    SelectedOrFullTimeRange(context, view_start, view_end);
+    if(!Actions(context).AddNote(time_ns, title, text, view_start, view_end,
+                                 JsonU64(args, "track_id", INVALID_UINT64_INDEX)))
+    {
+        return DoneResult("Could not pin a note on this trace.", "Note failed");
+    }
+    return DoneResult("Pinned a note titled \"" + title + "\" on the timeline.",
+                      "Left a note");
+}
+
+// Implements the switch_tab tool.
+AssistantToolStartResult
+ToolSwitchTab(const AssistantToolContext& context, const jt::Json& args,
+              const std::string&)
+{
+    OptiqActions                   actions       = Actions(context);
+    const std::vector<std::string> trace_tabs    = actions.ListTabs();
+    const std::vector<std::string> analysis_tabs = actions.ListAnalysisTabs();
+
+    std::ostringstream out;
+    out << "trace_tabs:";
+    for(const std::string& tab : trace_tabs)
+    {
+        out << "\n  " << tab;
+    }
+    out << "\nactive_trace_tab: " << actions.ActiveTab() << "\n";
+    if(!analysis_tabs.empty())
+    {
+        out << "details_panel_tabs:";
+        for(const std::string& tab : analysis_tabs)
+        {
+            out << "\n  " << tab;
+        }
+        out << "\nactive_details_tab: " << actions.ActiveAnalysisTab() << "\n";
+    }
+
+    const std::string name = JsonUtils::GetString(args, "name", "");
+    if(name.empty())
+    {
+        return DoneResult(out.str(), "Listed tabs");
+    }
+
+    // Details tabs are the ones people usually mean by name, so try those
+    // first; trace tabs are file names and rarely collide.
+    if(actions.SelectAnalysisTab(name))
+    {
+        return DoneResult("Opened the " + actions.ActiveAnalysisTab() +
+                              " tab in the details panel.",
+                          "Switched tabs");
+    }
+    if(actions.SelectTab(name))
+    {
+        return DoneResult("Switched to trace " + actions.ActiveTab() + ".",
+                          "Switched tabs");
+    }
+
+    out << "Nothing matches \"" << name << "\".\n";
+    return DoneResult(out.str(), "No matching tab");
+}
+
+// Implements the flow_arrows tool.
+AssistantToolStartResult
+ToolFlowArrows(const AssistantToolContext& context, const jt::Json& args,
+               const std::string&)
+{
+    OptiqActions actions = Actions(context);
+    const bool   visible = JsonUtils::GetBool(args, "visible", true);
+    if(!actions.SetFlowArrowsVisible(visible))
+    {
+        return DoneResult("Flow arrows are only available on a system trace.",
+                          "Not available");
+    }
+
+    const std::string style = to_lower_copy(JsonUtils::GetString(args, "style", ""));
+    if(style == "chain" || style == "fan")
+    {
+        actions.SetFlowRenderChained(style == "chain");
+    }
+
+    std::string message = visible ? "Flow arrows are on." : "Flow arrows are off.";
+    if(!style.empty())
+    {
+        message += " Style set to " + style + ".";
+    }
+    if(visible)
+    {
+        message +=
+            " They only draw for a selected event, so call goto with an "
+            "event_uuid and track_id to see them.";
+    }
+    return DoneResult(message, visible ? "Flow arrows on" : "Flow arrows off");
+}
+
+// Implements the trace_overview tool.
+AssistantToolStartResult
+ToolTraceOverview(const AssistantToolContext& context, const jt::Json& args,
+                  const std::string&)
+{
+    if(context.is_compute)
+    {
+        return DoneResult(
+            "Compute traces have no timeline. Call get_summary instead.",
+            "Wrong trace kind");
+    }
+    const int32_t requested_bins = JsonUtils::GetInt(
+        args, "bins", static_cast<int32_t>(ASSISTANT_OVERVIEW_BINS));
+    const size_t bins =
+        requested_bins <= 0
+            ? ASSISTANT_OVERVIEW_BINS
+            : std::min(static_cast<size_t>(requested_bins), ASSISTANT_MAX_ROW_LIMIT);
+    const uint64_t track_id = JsonU64(args, "track_id", INVALID_UINT64_INDEX);
+
+    AssistantToolStartResult result =
+        DoneResult(FormatTraceOverview(context, bins, track_id),
+                   "Read the timeline overview");
+    result.chart          = true;
+    result.chart_track_id = track_id;
+    return result;
+}
+
+// Implements the get_summary tool.
+AssistantToolStartResult
+ToolGetSummary(const AssistantToolContext& context, const jt::Json& args,
+               const std::string&)
+{
+    if(context.is_compute)
+    {
+        return DoneResult(FormatComputeKernels(context, ASSISTANT_TOP_KERNEL_LIMIT),
+                          "Read compute kernel stats");
+    }
+    const AssistantFetchState summary_fetch =
+        TableFetch(AssistantFetchKind::kSummary, TableType::kSummaryKernelTable,
+                   ASSISTANT_DEFAULT_ROW_LIMIT);
+    if(context.data_provider->IsRequestPending(DataProvider::SUMMARY_REQUEST_ID))
+    {
+        return PendingResult(DataProvider::SUMMARY_REQUEST_ID, summary_fetch,
+                             "Loading summary...", false);
+    }
+    const SummaryInfo::GPUMetrics& gpu =
+        context.data_provider->DataModel().GetSummary().GetSummaryData().gpu;
+    if(!gpu.top_kernels.empty() || gpu.kernel_exec_time_total > 0.0)
+    {
+        return DoneResult(FormatSystemSummary(context), "Read summary");
+    }
+    if(context.data_provider->FetchSummary())
+    {
+        return PendingResult(DataProvider::SUMMARY_REQUEST_ID, summary_fetch,
+                             "Loading summary...", true);
+    }
+    if(context.data_provider->IsRequestPending(DataProvider::SUMMARY_REQUEST_ID))
+    {
+        return PendingResult(DataProvider::SUMMARY_REQUEST_ID, summary_fetch,
+                             "Loading summary...", false);
+    }
+    // The fetch was refused and nothing is in flight, so the summary really
+    // is empty. Formatting it anyway would hand the model a page of zeroes
+    // that reads like a measurement.
+    return DoneResult("The GPU summary could not be loaded for this trace. "
+                      "Use top_events or track_events instead.",
+                      "Summary unavailable");
+}
+
+// Implements the list_tracks tool.
+AssistantToolStartResult
+ToolListTracks(const AssistantToolContext& context, const jt::Json& args,
+               const std::string&)
+{
+    if(context.is_compute)
+    {
+        return DoneResult("Compute traces have no timeline tracks. Use kernel_metrics.",
+                          "No timeline tracks");
+    }
+    const std::vector<const TrackInfo*> tracks =
+        context.data_provider->DataModel().GetTimeline().GetTrackList();
+    std::ostringstream out;
+    size_t             shown = 0;
+    for(const TrackInfo* track : tracks)
+    {
+        if(track == nullptr)
+        {
+            continue;
+        }
+        if(track->topology.type != TrackInfo::Queue &&
+           track->topology.type != TrackInfo::Stream &&
+           track->topology.type != TrackInfo::InstrumentedThread &&
+           track->topology.type != TrackInfo::SampledThread &&
+           track->topology.type != TrackInfo::Counter)
+        {
+            continue;
+        }
+        out << "track_id=" << track->id << " type=" << TrackTypeName(track->topology.type)
+            << " reads_with="
+            << (track->track_type == kRPVControllerTrackTypeEvents ? "track_events"
+                                                                  : "track_samples")
+            << " name=\""
+            << context.data_provider->DataModel().BuildTrackName(track->id) << "\"";
+        out << " ops=";
+        bool first = true;
+        AppendOpName(out, first, kRocProfVisDmOperationDispatch, *track);
+        AppendOpName(out, first, kRocProfVisDmOperationMemoryCopy, *track);
+        AppendOpName(out, first, kRocProfVisDmOperationMemoryAllocate, *track);
+        AppendOpName(out, first, kRocProfVisDmOperationLaunch, *track);
+        AppendOpName(out, first, kRocProfVisDmOperationLaunchSample, *track);
+        if(first)
+        {
+            out << "-";
+        }
+        out << "\n";
+        ++shown;
+        if(shown >= ASSISTANT_MAX_LIST_TRACKS)
+        {
+            out << "... additional tracks omitted\n";
+            break;
         }
     }
-    return "Using " + tool_name + "...";
+    if(shown == 0)
+    {
+        return DoneResult("No queue/thread tracks found.", "No tracks");
+    }
+    return DoneResult(TrimResult(out.str()), "Listed tracks");
 }
+
+// Implements the goto tool.
+AssistantToolStartResult
+ToolGoto(const AssistantToolContext& context, const jt::Json& args,
+         const std::string&)
+{
+    const double start_ns =
+        JsonUtils::GetDouble(args, "start_ns",
+                             JsonUtils::GetDouble(args, "start", -1.0));
+    const double end_ns =
+        JsonUtils::GetDouble(args, "end_ns", JsonUtils::GetDouble(args, "end", -1.0));
+    const uint64_t track_id =
+        JsonU64(args, "track_id", TimelineSelection::INVALID_SELECTION_ID);
+    const uint64_t event_uuid =
+        JsonU64(args, "event_uuid",
+                JsonU64(args, "event_id", TimelineSelection::INVALID_SELECTION_ID));
+    const std::string kernel_name = JsonUtils::GetString(args, "kernel_name", "");
+
+    if(context.is_compute)
+    {
+        if(context.compute_selection == nullptr)
+        {
+            return DoneResult("Compute selection is not available.", "goto failed");
+        }
+        ComputeDataModel& model = context.data_provider->ComputeModel();
+        uint32_t workload_id    = context.compute_selection->GetSelectedWorkload();
+        const WorkloadInfo* workload = model.GetWorkload(workload_id);
+        if(workload == nullptr && !model.GetWorkloadList().empty())
+        {
+            workload = model.GetWorkloadList().front();
+        }
+        if(workload == nullptr)
+        {
+            return DoneResult("No compute workload is loaded.", "goto failed");
+        }
+        const uint32_t kernel_id = static_cast<uint32_t>(
+            JsonU64(args, "kernel_id", ComputeSelection::INVALID_SELECTION_ID));
+        const KernelInfo* kernel = FindComputeKernel(*workload, kernel_name, kernel_id);
+        if(kernel == nullptr)
+        {
+            return DoneResult("Could not find that kernel to select.", "goto failed");
+        }
+        Actions(context).SelectKernel(kernel->id);
+        return DoneResult(std::string("Selected compute kernel ") + kernel->name,
+                          "Selected kernel");
+    }
+
+    // isfinite is doing real work here: NaN fails all three comparisons
+    // below, so without it a NaN bound would be accepted as a valid range.
+    if(!std::isfinite(start_ns) || !std::isfinite(end_ns) || start_ns < 0.0 ||
+       end_ns <= start_ns)
+    {
+        return DoneResult(
+            "goto needs start_ns and end_ns (end > start) from kernel_instances "
+            "or the selected range.",
+            "goto failed");
+    }
+    if(context.timeline_selection == nullptr)
+    {
+        return DoneResult("Timeline selection is not available.", "goto failed");
+    }
+
+    OptiqActions actions = Actions(context);
+    actions.SelectRange(start_ns, end_ns);
+    const bool zoom = JsonUtils::GetBool(args, "zoom", false);
+    if(zoom)
+    {
+        actions.ZoomToRange(start_ns, end_ns);
+    }
+
+    // The model may name one event via track_id/event_uuid or several via
+    // events[]; both shapes collapse to this list.
+    std::vector<std::pair<uint64_t, uint64_t>> targets;
+    if(event_uuid != TimelineSelection::INVALID_SELECTION_ID)
+    {
+        targets.emplace_back(track_id, event_uuid);
+    }
+    jt::Json& mutable_args = const_cast<jt::Json&>(args);
+    if(mutable_args.contains("events") && mutable_args["events"].isArray())
+    {
+        std::vector<jt::Json>& entries = mutable_args["events"].getArray();
+        std::string            length_error;
+        if(!CheckArrayLength(entries, "events", length_error))
+        {
+            return DoneResult(length_error, "goto failed");
+        }
+        for(jt::Json& entry : entries)
+        {
+            if(targets.size() >= ASSISTANT_MAX_HIGHLIGHTS)
+            {
+                break;
+            }
+            if(!entry.isObject())
+            {
+                continue;
+            }
+            const uint64_t entry_uuid =
+                JsonU64(entry, "event_uuid",
+                        JsonU64(entry, "__uuid",
+                                TimelineSelection::INVALID_SELECTION_ID));
+            if(entry_uuid != TimelineSelection::INVALID_SELECTION_ID)
+            {
+                targets.emplace_back(JsonU64(entry, "track_id", track_id), entry_uuid);
+            }
+        }
+    }
+
+    // Highlighting comes before navigating, so the view is framed on an
+    // event that actually took rather than on whichever the model listed
+    // first. Clicking that one is also what makes the trace view load its
+    // details, flow arrows, and call stack - NavigateToEvent only frames.
+    size_t                        highlighted = 0;
+    std::pair<uint64_t, uint64_t> first_highlighted(0, 0);
+    if(!targets.empty())
+    {
+        actions.ClearHighlights();
+        for(const std::pair<uint64_t, uint64_t>& target : targets)
+        {
+            if(!actions.HighlightEvent(target.first, target.second))
+            {
+                continue;
+            }
+            if(highlighted == 0)
+            {
+                first_highlighted = target;
+                actions.ClickEvent(target.first, target.second);
+                actions.ScrollToTrack(target.first);
+            }
+            ++highlighted;
+        }
+    }
+
+    if(highlighted > 0)
+    {
+        actions.NavigateToEvent(first_highlighted.first, first_highlighted.second,
+                                start_ns, end_ns - start_ns);
+    }
+    else
+    {
+        actions.ShowRange(start_ns, end_ns);
+    }
+
+    SettingsManager& settings    = SettingsManager::GetInstance();
+    const TimeFormat time_format = settings.GetUserSettings().unit_settings.time_format;
+    std::ostringstream out;
+    out << "Moved the timeline to "
+        << nanosecond_to_formatted_str(start_ns, time_format, true) << " .. "
+        << nanosecond_to_formatted_str(end_ns, time_format, true);
+    if(zoom)
+    {
+        out << ", zoomed in";
+    }
+    if(highlighted > 0)
+    {
+        out << " and highlighted " << highlighted
+            << " event(s) there, with flow arrows expanded on the first";
+    }
+    else if(!targets.empty())
+    {
+        // Saying "moved the view" alone would let the model believe it had
+        // selected events it never selected, and report that to the user.
+        out << ", but none of the " << targets.size()
+            << " event(s) named could be selected - check that __uuid and "
+               "__trackId came from a table read on this trace";
+    }
+    out << ".";
+    return DoneResult(out.str(), "Moved the view");
+}
+
+// Implements the top_events tool.
+AssistantToolStartResult
+ToolTopEvents(const AssistantToolContext& context, const jt::Json& args,
+              const std::string& tool_name)
+{
+    if(context.is_compute)
+    {
+        return DoneResult(
+            "top_events is for system traces. Use get_summary / kernel_metrics.",
+            "Wrong trace kind");
+    }
+    const std::string category = JsonUtils::GetString(args, "category", "dispatch");
+    const TopEventsSpec* spec  = FindTopEventsSpec(category);
+    if(spec == nullptr)
+    {
+        return DoneResult(
+            "Unknown category. Use dispatch, memory_copy, memory_alloc, "
+            "instrumented, or sampled.",
+            "Bad category");
+    }
+    const size_t limit =
+        ClampRowLimit(JsonUtils::GetInt(args, "limit",
+                                        static_cast<int32_t>(ASSISTANT_DEFAULT_ROW_LIMIT)));
+
+    std::string track_error;
+    const std::vector<uint64_t> tracks =
+        TracksFromArgs(context, args, spec->op, true, track_error);
+    if(!track_error.empty())
+    {
+        return DoneResult(track_error, "Bad track_ids");
+    }
+    if(tracks.empty())
+    {
+        return DoneResult(
+            std::string("No tracks contain ") + spec->key +
+                " events. Call list_tracks or try another category.",
+            "No matching tracks");
+    }
+
+    std::string       query_error;
+    const std::string where    = BuildAssistantWhereClause(args, query_error);
+    const std::string group_by = AssistantGroupByFromArgs(args, query_error);
+    if(!query_error.empty())
+    {
+        return DoneResult(query_error, "Bad query argument");
+    }
+
+    double start_ns = 0.0;
+    double end_ns   = 0.0;
+    TimeRangeFromArgs(context, args, start_ns, end_ns);
+
+    const AssistantFetchState fetch =
+        TableFetch(AssistantFetchKind::kTopEvents, spec->table_type, limit);
+    if(context.data_provider->IsRequestPending(spec->request_id))
+    {
+        return PendingResult(spec->request_id, fetch,
+                             AssistantToolStatusLabel(tool_name), false);
+    }
+
+    const TablesModel& tables =
+        context.data_provider->DataModel().GetAnalysis().GetTables();
+    const bool queued = context.data_provider->FetchTable(TrackTableRequestParams(
+        spec->controller_type, tracks, start_ns, end_ns, where.c_str(), "",
+        group_by.c_str(), "", OffsetFromArgs(args),
+        static_cast<uint64_t>(limit),
+        ResolveAssistantSortColumn(tables, spec->table_type,
+                          JsonUtils::GetString(args, "sort_by", ""),
+                          ASSISTANT_DURATION_COLUMN),
+        AssistantSortOrderFromArgs(args, kRPVControllerSortOrderDescending)));
+    if(!queued && !context.data_provider->IsRequestPending(spec->request_id))
+    {
+        return DoneResult("Could not queue the top_events fetch.", "Fetch failed");
+    }
+    return PendingResult(spec->request_id, fetch,
+                         AssistantToolStatusLabel(tool_name), queued);
+}
+
+// Implements the kernel_instances tool.
+AssistantToolStartResult
+ToolKernelInstances(const AssistantToolContext& context, const jt::Json& args,
+                    const std::string& tool_name)
+{
+    if(context.is_compute)
+    {
+        return DoneResult(
+            "kernel_instances is for system traces. Use kernel_metrics on compute traces.",
+            "Wrong trace kind");
+    }
+    const std::string kernel_name = JsonUtils::GetString(args, "kernel_name", "");
+    if(kernel_name.empty())
+    {
+        return DoneResult("kernel_instances needs kernel_name.", "Missing kernel_name");
+    }
+    const size_t limit =
+        ClampRowLimit(JsonUtils::GetInt(args, "limit",
+                                        static_cast<int32_t>(ASSISTANT_DEFAULT_ROW_LIMIT)));
+
+    std::string       query_error;
+    const std::string where = BuildAssistantWhereClause(args, query_error);
+    if(!query_error.empty())
+    {
+        return DoneResult(query_error, "Bad query argument");
+    }
+
+    double start_ns = 0.0;
+    double end_ns   = 0.0;
+    TimeRangeFromArgs(context, args, start_ns, end_ns);
+
+    const uint64_t request_id = DataProvider::SUMMARY_KERNEL_INSTANCE_TABLE_REQUEST_ID;
+    const AssistantFetchState fetch =
+        TableFetch(AssistantFetchKind::kKernelInstances, TableType::kSummaryKernelTable,
+                   limit);
+    if(context.data_provider->IsRequestPending(request_id))
+    {
+        return PendingResult(request_id, fetch, AssistantToolStatusLabel(tool_name),
+                             false);
+    }
+
+    const TablesModel& tables = context.data_provider->DataModel().GetTables();
+    const bool queued = context.data_provider->FetchTable(EventSearchRequestParams(
+        kRPVControllerTableTypeSummaryKernelInstances,
+        { kRocProfVisDmOperationDispatch }, start_ns, end_ns, where.c_str(), false,
+        false, false, { kernel_name }, OffsetFromArgs(args),
+        static_cast<uint64_t>(limit),
+        ResolveAssistantSortColumn(tables, TableType::kSummaryKernelTable,
+                          JsonUtils::GetString(args, "sort_by", ""), 0),
+        AssistantSortOrderFromArgs(args, kRPVControllerSortOrderAscending)));
+    if(!queued && !context.data_provider->IsRequestPending(request_id))
+    {
+        return DoneResult("Could not queue the kernel_instances fetch.", "Fetch failed");
+    }
+    return PendingResult(request_id, fetch, AssistantToolStatusLabel(tool_name),
+                         queued);
+}
+
+// Implements the kernel_metrics tool.
+AssistantToolStartResult
+ToolKernelMetrics(const AssistantToolContext& context, const jt::Json& args,
+                  const std::string& tool_name)
+{
+    const std::string kernel_name = JsonUtils::GetString(args, "kernel_name", "");
+    const std::string metric_name = JsonUtils::GetString(args, "metric_name", "");
+    if(!context.is_compute)
+    {
+        std::ostringstream out;
+        const std::vector<SummaryInfo::KernelMetrics>& kernels =
+            context.data_provider->DataModel().GetSummary().GetSummaryData().gpu.top_kernels;
+        bool found = false;
+        for(const SummaryInfo::KernelMetrics& kernel : kernels)
+        {
+            if(kernel_name.empty() || kernel.name == kernel_name ||
+               to_lower_copy(kernel.name).find(to_lower_copy(kernel_name)) !=
+                   std::string::npos)
+            {
+                out << "name=" << kernel.name << " calls=" << kernel.invocations
+                    << " sum_ns=" << kernel.exec_time_sum
+                    << " min_ns=" << kernel.exec_time_min
+                    << " max_ns=" << kernel.exec_time_max
+                    << " pct=" << kernel.exec_time_pct << "\n";
+                found = true;
+                if(!kernel_name.empty())
+                {
+                    break;
+                }
+            }
+        }
+        if(!found)
+        {
+            out << "Kernel not in the current summary. Call get_summary first.\n";
+        }
+        return DoneResult(out.str(), "Read kernel summary stats");
+    }
+
+    ComputeDataModel& model = context.data_provider->ComputeModel();
+    uint32_t workload_id    = ComputeSelection::INVALID_SELECTION_ID;
+    if(context.compute_selection != nullptr)
+    {
+        workload_id = context.compute_selection->GetSelectedWorkload();
+    }
+    const WorkloadInfo* workload = model.GetWorkload(workload_id);
+    if(workload == nullptr && !model.GetWorkloadList().empty())
+    {
+        workload = model.GetWorkloadList().front();
+    }
+    if(workload == nullptr)
+    {
+        return DoneResult("No compute workload is loaded.", "No workload");
+    }
+
+    uint32_t selected_kernel = ComputeSelection::INVALID_SELECTION_ID;
+    if(context.compute_selection != nullptr)
+    {
+        selected_kernel = context.compute_selection->GetSelectedKernel();
+    }
+    const uint32_t kernel_id_arg = static_cast<uint32_t>(
+        JsonU64(args, "kernel_id", selected_kernel));
+    const KernelInfo* kernel =
+        FindComputeKernel(*workload, kernel_name, kernel_id_arg);
+    if(kernel == nullptr && selected_kernel != ComputeSelection::INVALID_SELECTION_ID)
+    {
+        kernel = FindComputeKernel(*workload, "", selected_kernel);
+    }
+    if(kernel == nullptr)
+    {
+        return DoneResult("Could not find that kernel. Call get_summary.",
+                          "Unknown kernel");
+    }
+
+    std::ostringstream out;
+    out << "kernel_id=" << kernel->id << " name=" << kernel->name << "\n";
+    out << "invocation_count=" << kernel->dispatch_metrics[KernelInfo::InvocationCount]
+        << "\n";
+    out << "duration_total_ns=" << kernel->dispatch_metrics[KernelInfo::DurationTotal]
+        << "\n";
+    out << "duration_min_ns=" << kernel->dispatch_metrics[KernelInfo::DurationMin]
+        << "\n";
+    out << "duration_max_ns=" << kernel->dispatch_metrics[KernelInfo::DurationMax]
+        << "\n";
+    out << "duration_mean_ns=" << kernel->dispatch_metrics[KernelInfo::DurationMean]
+        << "\n";
+    out << "duration_median_ns="
+        << kernel->dispatch_metrics[KernelInfo::DurationMedian] << "\n";
+
+    if(metric_name.empty())
+    {
+        out << "available_metrics:\n";
+        size_t listed = 0;
+        for(const AvailableMetrics::Entry& entry : workload->available_metrics.list)
+        {
+            out << "  " << entry.name;
+            if(!entry.unit.empty())
+            {
+                out << " (" << entry.unit << ")";
+            }
+            out << "\n";
+            ++listed;
+            if(listed >= ASSISTANT_MAX_METRIC_NAMES)
+            {
+                out << "  ... more metrics omitted; pass metric_name to fetch a subset\n";
+                break;
+            }
+        }
+        return DoneResult(TrimResult(out.str()), "Read kernel dispatch stats");
+    }
+
+    const std::string needle = to_lower_copy(metric_name);
+    std::vector<MetricsRequestParams::MetricID> metric_ids;
+    for(const AvailableMetrics::Entry& entry : workload->available_metrics.list)
+    {
+        if(to_lower_copy(entry.name).find(needle) == std::string::npos)
+        {
+            continue;
+        }
+        MetricsRequestParams::MetricID id;
+        id.category_id = entry.category_id;
+        id.table_id    = entry.table_id;
+        id.entry_id    = entry.id;
+        metric_ids.push_back(id);
+        if(metric_ids.size() >= ASSISTANT_MAX_METRIC_FETCH)
+        {
+            break;
+        }
+    }
+    if(metric_ids.empty())
+    {
+        out << "No available metric name contains \"" << metric_name << "\".\n";
+        return DoneResult(TrimResult(out.str()), "Metric not found");
+    }
+
+    model.ClearKernelMetricValues(context.metrics_client_id);
+    const uint64_t request_id = RequestIdBuilder::MakeClientRequestId(
+        RequestType::kFetchMetrics, context.metrics_client_id);
+    AssistantFetchState metrics_fetch =
+        TableFetch(AssistantFetchKind::kMetrics, TableType::kSummaryKernelTable,
+                   ASSISTANT_DEFAULT_ROW_LIMIT);
+    metrics_fetch.kernel_id = kernel->id;
+    if(context.data_provider->IsRequestPending(request_id))
+    {
+        return PendingResult(request_id, metrics_fetch,
+                             AssistantToolStatusLabel(tool_name), false);
+    }
+    const bool queued = context.data_provider->FetchMetrics(MetricsRequestParams(
+        workload->id, { kernel->id }, metric_ids, context.metrics_client_id));
+    if(!queued && !context.data_provider->IsRequestPending(request_id))
+    {
+        out << "Could not queue the metric fetch.\n";
+        return DoneResult(TrimResult(out.str()), "Fetch failed");
+    }
+    AssistantToolStartResult pending = PendingResult(
+        request_id, metrics_fetch, AssistantToolStatusLabel(tool_name), queued);
+    pending.content = out.str();
+    return pending;
+}
+
+// Implements track_events and track_samples, which differ only in which table
+// they read; tool_name is what tells them apart.
+AssistantToolStartResult
+ToolTrackRows(const AssistantToolContext& context, const jt::Json& args,
+                const std::string& tool_name)
+{
+    if(context.is_compute)
+    {
+        return DoneResult(
+            "Compute traces have no timeline tracks. Use kernel_metrics.",
+            "Wrong trace kind");
+    }
+
+    const bool      events   = tool_name == "track_events";
+    const TableType type     = events ? TableType::kEventTable : TableType::kSampleTable;
+    const uint64_t request_id = events ? DataProvider::EVENT_TABLE_REQUEST_ID
+                                       : DataProvider::SAMPLE_TABLE_REQUEST_ID;
+    const rocprofvis_controller_track_type_t wanted =
+        events ? kRPVControllerTrackTypeEvents : kRPVControllerTrackTypeSamples;
+
+    std::string           track_error;
+    std::vector<uint64_t> tracks = TracksFromArgs(
+        context, args, kRocProfVisDmOperationDispatch, false, track_error);
+    if(!track_error.empty())
+    {
+        return DoneResult(track_error, "Bad track_ids");
+    }
+    if(tracks.empty())
+    {
+        tracks = TracksOfType(context, wanted);
+    }
+    else
+    {
+        tracks = KeepTracksOfType(context, tracks, wanted);
+        if(tracks.empty())
+        {
+            return DoneResult(
+                std::string("Those track_ids are not ") +
+                    (events ? "event tracks. Use track_samples for counter tracks."
+                            : "counter tracks. Use track_events for queue, stream, "
+                              "and thread tracks."),
+                "Wrong track kind");
+        }
+    }
+    if(tracks.empty())
+    {
+        return DoneResult(std::string("No ") +
+                              (events ? "event" : "counter") +
+                              " tracks found. Call list_tracks.",
+                          "No matching tracks");
+    }
+
+    std::string       query_error;
+    const std::string where    = BuildAssistantWhereClause(args, query_error);
+    const std::string group_by = AssistantGroupByFromArgs(args, query_error);
+    if(!query_error.empty())
+    {
+        return DoneResult(query_error, "Bad query argument");
+    }
+
+    const size_t limit =
+        ClampRowLimit(JsonUtils::GetInt(args, "limit",
+                                        static_cast<int32_t>(ASSISTANT_DEFAULT_ROW_LIMIT)));
+    double start_ns = 0.0;
+    double end_ns   = 0.0;
+    TimeRangeFromArgs(context, args, start_ns, end_ns);
+
+    const AssistantFetchState fetch =
+        TableFetch(AssistantFetchKind::kDataTable, type, limit);
+    if(context.data_provider->IsRequestPending(request_id))
+    {
+        return PendingResult(request_id, fetch, AssistantToolStatusLabel(tool_name),
+                             false);
+    }
+
+    const TablesModel& tables = context.data_provider->DataModel().GetTables();
+    const bool queued = context.data_provider->FetchTable(TrackTableRequestParams(
+        events ? kRPVControllerTableTypeEvents : kRPVControllerTableTypeSamples,
+        tracks, start_ns, end_ns, where.c_str(), "", group_by.c_str(), "",
+        OffsetFromArgs(args), static_cast<uint64_t>(limit),
+        ResolveAssistantSortColumn(tables, type, JsonUtils::GetString(args, "sort_by", ""), 0),
+        AssistantSortOrderFromArgs(args, kRPVControllerSortOrderAscending)));
+    if(!queued && !context.data_provider->IsRequestPending(request_id))
+    {
+        return DoneResult("Could not queue the " + tool_name + " fetch.",
+                          "Fetch failed");
+    }
+    return PendingResult(request_id, fetch, AssistantToolStatusLabel(tool_name),
+                         queued);
+}
+
+// Implements the event_details tool.
+AssistantToolStartResult
+ToolEventDetails(const AssistantToolContext& context, const jt::Json& args,
+                 const std::string& tool_name)
+{
+    if(context.is_compute)
+    {
+        return DoneResult("event_details is for system traces.", "Wrong trace kind");
+    }
+    const uint64_t event_uuid =
+        JsonU64(args, "event_uuid", JsonU64(args, "event_id", 0));
+    if(event_uuid == 0)
+    {
+        return DoneResult("event_details needs event_uuid from the __uuid column.",
+                          "Missing event_uuid");
+    }
+
+    const std::vector<uint64_t> request_ids = {
+        DataProvider::EVENT_EXTENDED_DATA_REQUEST_ID,
+        DataProvider::EVENT_FLOW_DATA_REQUEST_ID,
+        DataProvider::EVENT_CALL_STACK_DATA_REQUEST_ID
+    };
+    // These request ids are shared with the event panel, and re-issuing one
+    // that is already in flight would drop its future on the floor.
+    for(uint64_t request_id : request_ids)
+    {
+        if(context.data_provider->IsRequestPending(request_id))
+        {
+            return DoneResult("Event details are already loading. Try again.",
+                              "Busy");
+        }
+    }
+
+    const uint64_t track_id = JsonU64(args, "track_id", INVALID_UINT64_INDEX);
+    // FetchEvent seeds name/duration from loaded track data and chains the
+    // extended-data fetch; without a track id only the fetches below run.
+    bool extended = false;
+    if(track_id != INVALID_UINT64_INDEX)
+    {
+        extended = context.data_provider->FetchEvent(track_id, event_uuid);
+    }
+    const bool flow  = context.data_provider->FetchEventFlowDetails(event_uuid);
+    const bool stack = context.data_provider->FetchEventCallStackData(event_uuid);
+    if(!extended && !flow && !stack)
+    {
+        return DoneResult("Could not load details for that event.", "Fetch failed");
+    }
+
+    AssistantFetchState fetch;
+    fetch.kind     = AssistantFetchKind::kEventDetails;
+    fetch.event_id = event_uuid;
+    fetch.track_id = track_id;
+    return PendingResult(request_ids, fetch, AssistantToolStatusLabel(tool_name),
+                         true);
+}
+
+// Implements the track_statistics tool.
+AssistantToolStartResult
+ToolTrackStatistics(const AssistantToolContext& context, const jt::Json& args,
+                    const std::string& tool_name)
+{
+    if(context.is_compute)
+    {
+        return DoneResult("track_statistics is for system traces.",
+                          "Wrong trace kind");
+    }
+    const uint64_t track_id = JsonU64(args, "track_id", INVALID_UINT64_INDEX);
+    const TrackInfo* track =
+        context.data_provider->DataModel().GetTimeline().GetTrack(track_id);
+    if(track == nullptr)
+    {
+        return DoneResult("Unknown track_id. Call list_tracks first.",
+                          "Unknown track");
+    }
+    if(track->topology.type != TrackInfo::Queue &&
+       track->topology.type != TrackInfo::Counter)
+    {
+        return DoneResult(
+            "track_statistics only works on Queue and Counter tracks. Use "
+            "track_events for thread and stream tracks.",
+            "Wrong track type");
+    }
+
+    const AnalysisTrackStatistics* stats =
+        context.data_provider->DataModel().GetAnalysis().RegisterTrack(*track);
+    if(stats == nullptr)
+    {
+        return DoneResult("Statistics are not available for that track.",
+                          "No statistics");
+    }
+
+    double start_ns = 0.0;
+    double end_ns   = 0.0;
+    TimeRangeFromArgs(context, args, start_ns, end_ns);
+
+    AssistantFetchState fetch;
+    fetch.kind     = AssistantFetchKind::kTrackStatistics;
+    fetch.track_id = track_id;
+
+    const uint64_t request_id = RequestIdBuilder::MakeTrackDataRequestId(
+        static_cast<uint32_t>(track_id), 0, 0,
+        RequestType::kFetchAnalysisTrackStatistics);
+    if(context.data_provider->IsRequestPending(request_id))
+    {
+        return PendingResult(request_id, fetch, AssistantToolStatusLabel(tool_name),
+                             false);
+    }
+
+    const bool queued = context.data_provider->FetchAnalysisTrackStatistics(
+        AnalysisTrackStatisticsRequestParams(track_id, start_ns, end_ns));
+    stats->state = queued ? AnalysisTrackStatistics::kRequested
+                          : AnalysisTrackStatistics::kPending;
+    if(!queued)
+    {
+        return DoneResult("Could not queue the track statistics fetch.",
+                          "Fetch failed");
+    }
+    return PendingResult(request_id, fetch, AssistantToolStatusLabel(tool_name), true);
+}
+
+// Implements the search_events tool.
+AssistantToolStartResult
+ToolSearchEvents(const AssistantToolContext& context, const jt::Json& args,
+                 const std::string& tool_name)
+{
+    if(context.is_compute)
+    {
+        return DoneResult(
+            "search_events is for system traces. Use kernel_metrics on compute "
+            "traces.",
+            "Wrong trace kind");
+    }
+
+    std::vector<std::string> terms = JsonUtils::GetStringArray(args, "terms");
+    const std::string        single = JsonUtils::GetString(args, "query", "");
+    if(terms.empty() && !single.empty())
+    {
+        terms.push_back(single);
+    }
+    if(terms.empty())
+    {
+        return DoneResult(
+            "search_events needs terms, e.g. {\"terms\":[\"gemm\"]}.",
+            "Missing terms");
+    }
+    if(terms.size() > ASSISTANT_MAX_SEARCH_TERMS)
+    {
+        terms.resize(ASSISTANT_MAX_SEARCH_TERMS);
+    }
+
+    std::vector<rocprofvis_dm_event_operation_t> ops;
+    for(const std::string& category : JsonUtils::GetStringArray(args, "categories"))
+    {
+        const TopEventsSpec* spec = FindTopEventsSpec(category);
+        if(spec == nullptr)
+        {
+            return DoneResult("Unknown category \"" + category +
+                                  "\". Use dispatch, memory_copy, memory_alloc, "
+                                  "instrumented, or sampled.",
+                              "Bad category");
+        }
+        // Only a handful of categories exist, so a repeat is the model
+        // being redundant rather than asking for more. Dropping duplicates
+        // also bounds this list however long the argument was.
+        if(std::find(ops.begin(), ops.end(), spec->op) == ops.end())
+        {
+            ops.push_back(spec->op);
+        }
+    }
+    if(ops.empty())
+    {
+        ops = { kRocProfVisDmOperationLaunch, kRocProfVisDmOperationDispatch,
+                kRocProfVisDmOperationMemoryCopy,
+                kRocProfVisDmOperationMemoryAllocate,
+                kRocProfVisDmOperationLaunchSample };
+    }
+
+    const bool contains =
+        to_lower_copy(JsonUtils::GetString(args, "match", "contains")) != "equals";
+    const bool any_term =
+        to_lower_copy(JsonUtils::GetString(args, "combine", "all")) == "any";
+    const bool include_category =
+        JsonUtils::GetBool(args, "include_category", false);
+
+    std::string       query_error;
+    const std::string where = BuildAssistantWhereClause(args, query_error);
+    if(!query_error.empty())
+    {
+        return DoneResult(query_error, "Bad query argument");
+    }
+
+    const size_t limit =
+        ClampRowLimit(JsonUtils::GetInt(args, "limit",
+                                        static_cast<int32_t>(ASSISTANT_DEFAULT_ROW_LIMIT)));
+    double start_ns = 0.0;
+    double end_ns   = 0.0;
+    TimeRangeFromArgs(context, args, start_ns, end_ns, false);
+
+    const uint64_t            request_id = DataProvider::EVENT_SEARCH_REQUEST_ID;
+    const AssistantFetchState fetch = TableFetch(
+        AssistantFetchKind::kDataTable, TableType::kEventSearchTable, limit);
+    if(context.data_provider->IsRequestPending(request_id))
+    {
+        return PendingResult(request_id, fetch, AssistantToolStatusLabel(tool_name),
+                             false);
+    }
+
+    const TablesModel& tables = context.data_provider->DataModel().GetTables();
+    const bool queued = context.data_provider->FetchTable(EventSearchRequestParams(
+        kRPVControllerTableTypeSearchResults, ops, start_ns, end_ns, where.c_str(),
+        contains, include_category, any_term, terms, OffsetFromArgs(args),
+        static_cast<uint64_t>(limit),
+        ResolveAssistantSortColumn(tables, TableType::kEventSearchTable,
+                          JsonUtils::GetString(args, "sort_by", ""), 1),
+        AssistantSortOrderFromArgs(args, kRPVControllerSortOrderAscending)));
+    if(!queued && !context.data_provider->IsRequestPending(request_id))
+    {
+        return DoneResult("Could not queue the search.", "Fetch failed");
+    }
+    return PendingResult(request_id, fetch, AssistantToolStatusLabel(tool_name),
+                         queued);
+}
+
+// Maps the name the model used to the body that implements it. Kept beside
+// the bodies so adding a tool is one entry and one function, and so a name
+// that is registered in the schema but missing here fails loudly.
+struct AssistantToolEntry
+{
+    const char*          name;
+    AssistantToolHandler handler;
+};
+
+const AssistantToolEntry ASSISTANT_TOOL_HANDLERS[] = {
+    { "offer_next_steps", ToolOfferNextSteps },
+    { "show_panel", ToolShowPanel },
+    { "reset_view", ToolResetView },
+    { "measure", ToolMeasure },
+    { "bookmark", ToolBookmark },
+    { "annotate", ToolAnnotate },
+    { "switch_tab", ToolSwitchTab },
+    { "flow_arrows", ToolFlowArrows },
+    { "trace_overview", ToolTraceOverview },
+    { "get_summary", ToolGetSummary },
+    { "list_tracks", ToolListTracks },
+    { "goto", ToolGoto },
+    { "top_events", ToolTopEvents },
+    { "kernel_instances", ToolKernelInstances },
+    { "kernel_metrics", ToolKernelMetrics },
+    { "track_events", ToolTrackRows },
+    { "track_samples", ToolTrackRows },
+    { "event_details", ToolEventDetails },
+    { "track_statistics", ToolTrackStatistics },
+    { "search_events", ToolSearchEvents },
+};
+
+}  // namespace
 
 // Dispatches one named tool call, returning either a finished result or a
 // pending fetch the panel polls until the rows land.
@@ -2047,30 +2627,6 @@ AssistantToolStartResult
 StartAssistantTool(const AssistantToolContext& context, const std::string& tool_name,
                    const std::string& arguments_json)
 {
-    // Next-step buttons are a UI side-effect, so they do not need an open trace.
-    if(tool_name == "offer_next_steps")
-    {
-        bool           args_ok = true;
-        const jt::Json args    = ParseArgsObject(arguments_json, args_ok);
-        if(!args_ok)
-        {
-            return DoneResult("Those arguments were not a JSON object, so none of them "
-                              "were read. Send arguments as one JSON object.",
-                              "Bad arguments");
-        }
-        return OfferNextStepsResult(args);
-    }
-
-    if(context.data_provider == nullptr)
-    {
-        return DoneResult("No trace is open.", "No trace");
-    }
-    if(context.data_provider->GetState() != ProviderState::kReady)
-    {
-        return DoneResult("The trace is still loading. Wait and try again.",
-                          "Trace not ready");
-    }
-
     bool           args_ok = true;
     const jt::Json args    = ParseArgsObject(arguments_json, args_ok);
     if(!args_ok)
@@ -2080,1075 +2636,31 @@ StartAssistantTool(const AssistantToolContext& context, const std::string& tool_
                           "Bad arguments");
     }
 
-    if(tool_name == "show_panel")
+    // Next-step buttons are a UI side-effect, so they are the one tool that
+    // does not need an open trace.
+    if(tool_name != "offer_next_steps")
     {
-        const std::string panel_name = JsonUtils::GetString(args, "panel", "");
-        const OptiqPanel  panel      = OptiqActions::PanelFromName(panel_name);
-        if(panel == OptiqPanel::kUnknown)
+        if(context.data_provider == nullptr)
         {
-            return DoneResult("Unknown panel \"" + panel_name + "\". Use one of: " +
-                                  OptiqActions::PanelNameList() + ".",
-                              "Unknown panel");
+            return DoneResult("No trace is open.", "No trace");
         }
-
-        const bool visible = JsonUtils::GetBool(args, "visible", true);
-        if(!Actions(context).ShowPanel(panel, visible))
+        if(context.data_provider->GetState() != ProviderState::kReady)
         {
-            return DoneResult(std::string("Could not change the ") +
-                                  OptiqActions::PanelName(panel) + " panel here.",
-                              "Panel unavailable");
+            return DoneResult("The trace is still loading. Wait and try again.",
+                              "Trace not ready");
         }
-        return DoneResult(std::string(visible ? "Opened the " : "Closed the ") +
-                              OptiqActions::PanelName(panel) + " panel.",
-                          visible ? "Opened a panel" : "Closed a panel");
     }
 
-    if(tool_name == "reset_view")
+    for(const AssistantToolEntry& entry : ASSISTANT_TOOL_HANDLERS)
     {
-        if(!Actions(context).ResetView())
+        if(tool_name == entry.name)
         {
-            return DoneResult("Reset view is only available on a system trace.",
-                              "Not available");
+            return entry.handler(context, args, tool_name);
         }
-        return DoneResult("Zoomed back out to the whole trace.", "Reset the view");
     }
 
-    if(tool_name == "measure")
-    {
-        OptiqActions actions = Actions(context);
-        if(JsonUtils::GetBool(args, "clear", false))
-        {
-            actions.ClearMeasurement();
-            return DoneResult("Cleared the measurement.", "Cleared measurement");
-        }
-
-        const double start_ns = JsonUtils::GetDouble(args, "start_ns", -1.0);
-        const double end_ns   = JsonUtils::GetDouble(args, "end_ns", -1.0);
-        if(start_ns < 0.0 || end_ns <= start_ns)
-        {
-            return DoneResult("measure needs start_ns and end_ns with end > start.",
-                              "Bad range");
-        }
-        if(!actions.MeasureRange(start_ns, end_ns))
-        {
-            return DoneResult("Could not measure on this trace.", "Not available");
-        }
-
-        SettingsManager& settings    = SettingsManager::GetInstance();
-        const TimeFormat time_format = settings.GetUserSettings().unit_settings.time_format;
-        return DoneResult(
-            "Measured " +
-                nanosecond_to_formatted_str(end_ns - start_ns, time_format, true) +
-                " on the timeline.",
-            "Measured a span");
-    }
-
-    if(tool_name == "bookmark")
-    {
-        OptiqActions      actions = Actions(context);
-        const std::string action  = ToLowerCopy(JsonUtils::GetString(args, "action", "list"));
-        const std::vector<int> slots = actions.ListBookmarks();
-
-        std::ostringstream saved;
-        saved << "saved_bookmarks:";
-        if(slots.empty())
-        {
-            saved << " none";
-        }
-        for(int slot : slots)
-        {
-            saved << " " << slot;
-        }
-        saved << "\n";
-
-        if(action == "list")
-        {
-            return DoneResult(saved.str(), "Listed bookmarks");
-        }
-
-        const int slot = static_cast<int>(JsonUtils::GetInt(args, "slot", -1));
-        if(slot < 0)
-        {
-            return DoneResult("bookmark needs a slot from 0 to 9.\n" + saved.str(),
-                              "Missing slot");
-        }
-        if(action == "save")
-        {
-            if(!actions.SaveBookmark(slot))
-            {
-                return DoneResult("Could not save that bookmark.", "Bookmark failed");
-            }
-            return DoneResult("Saved the current view to bookmark " +
-                                  std::to_string(slot) + ".",
-                              "Saved a bookmark");
-        }
-        if(action == "goto")
-        {
-            if(!actions.GotoBookmark(slot))
-            {
-                return DoneResult("Bookmark " + std::to_string(slot) +
-                                      " is empty.\n" + saved.str(),
-                                  "No such bookmark");
-            }
-            return DoneResult("Jumped to bookmark " + std::to_string(slot) + ".",
-                              "Used a bookmark");
-        }
-        if(action == "remove")
-        {
-            if(!actions.RemoveBookmark(slot))
-            {
-                return DoneResult("Bookmark " + std::to_string(slot) + " is empty.",
-                                  "No such bookmark");
-            }
-            return DoneResult("Removed bookmark " + std::to_string(slot) + ".",
-                              "Removed a bookmark");
-        }
-        return DoneResult("Unknown action. Use save, goto, remove, or list.",
-                          "Bad action");
-    }
-
-    if(tool_name == "annotate")
-    {
-        if(context.is_compute)
-        {
-            return DoneResult("Notes are a timeline feature, so system traces only.",
-                              "Wrong trace kind");
-        }
-        const double      time_ns = JsonUtils::GetDouble(args, "time_ns", -1.0);
-        const std::string title   = JsonUtils::GetString(args, "title", "");
-        const std::string text    = JsonUtils::GetString(args, "text", "");
-        if(time_ns < 0.0 || title.empty() || text.empty())
-        {
-            return DoneResult("annotate needs time_ns, title, and text.",
-                              "Missing arguments");
-        }
-
-        double view_start = 0.0;
-        double view_end   = 0.0;
-        SelectedOrFullTimeRange(context, view_start, view_end);
-        if(!Actions(context).AddNote(time_ns, title, text, view_start, view_end,
-                                     JsonU64(args, "track_id", INVALID_UINT64_INDEX)))
-        {
-            return DoneResult("Could not pin a note on this trace.", "Note failed");
-        }
-        return DoneResult("Pinned a note titled \"" + title + "\" on the timeline.",
-                          "Left a note");
-    }
-
-    if(tool_name == "switch_tab")
-    {
-        OptiqActions                   actions       = Actions(context);
-        const std::vector<std::string> trace_tabs    = actions.ListTabs();
-        const std::vector<std::string> analysis_tabs = actions.ListAnalysisTabs();
-
-        std::ostringstream out;
-        out << "trace_tabs:";
-        for(const std::string& tab : trace_tabs)
-        {
-            out << "\n  " << tab;
-        }
-        out << "\nactive_trace_tab: " << actions.ActiveTab() << "\n";
-        if(!analysis_tabs.empty())
-        {
-            out << "details_panel_tabs:";
-            for(const std::string& tab : analysis_tabs)
-            {
-                out << "\n  " << tab;
-            }
-            out << "\nactive_details_tab: " << actions.ActiveAnalysisTab() << "\n";
-        }
-
-        const std::string name = JsonUtils::GetString(args, "name", "");
-        if(name.empty())
-        {
-            return DoneResult(out.str(), "Listed tabs");
-        }
-
-        // Details tabs are the ones people usually mean by name, so try those
-        // first; trace tabs are file names and rarely collide.
-        if(actions.SelectAnalysisTab(name))
-        {
-            return DoneResult("Opened the " + actions.ActiveAnalysisTab() +
-                                  " tab in the details panel.",
-                              "Switched tabs");
-        }
-        if(actions.SelectTab(name))
-        {
-            return DoneResult("Switched to trace " + actions.ActiveTab() + ".",
-                              "Switched tabs");
-        }
-
-        out << "Nothing matches \"" << name << "\".\n";
-        return DoneResult(out.str(), "No matching tab");
-    }
-
-    if(tool_name == "flow_arrows")
-    {
-        OptiqActions actions = Actions(context);
-        const bool   visible = JsonUtils::GetBool(args, "visible", true);
-        if(!actions.SetFlowArrowsVisible(visible))
-        {
-            return DoneResult("Flow arrows are only available on a system trace.",
-                              "Not available");
-        }
-
-        const std::string style = ToLowerCopy(JsonUtils::GetString(args, "style", ""));
-        if(style == "chain" || style == "fan")
-        {
-            actions.SetFlowRenderChained(style == "chain");
-        }
-
-        std::string message = visible ? "Flow arrows are on." : "Flow arrows are off.";
-        if(!style.empty())
-        {
-            message += " Style set to " + style + ".";
-        }
-        if(visible)
-        {
-            message +=
-                " They only draw for a selected event, so call goto with an "
-                "event_uuid and track_id to see them.";
-        }
-        return DoneResult(message, visible ? "Flow arrows on" : "Flow arrows off");
-    }
-
-    if(tool_name == "trace_overview")
-    {
-        if(context.is_compute)
-        {
-            return DoneResult(
-                "Compute traces have no timeline. Call get_summary instead.",
-                "Wrong trace kind");
-        }
-        const int32_t requested_bins = JsonUtils::GetInt(
-            args, "bins", static_cast<int32_t>(ASSISTANT_OVERVIEW_BINS));
-        const size_t bins =
-            requested_bins <= 0
-                ? ASSISTANT_OVERVIEW_BINS
-                : std::min(static_cast<size_t>(requested_bins), ASSISTANT_MAX_ROW_LIMIT);
-        const uint64_t track_id = JsonU64(args, "track_id", INVALID_UINT64_INDEX);
-
-        AssistantToolStartResult result =
-            DoneResult(FormatTraceOverview(context, bins, track_id),
-                       "Read the timeline overview");
-        result.chart          = true;
-        result.chart_track_id = track_id;
-        return result;
-    }
-
-    if(tool_name == "get_summary")
-    {
-        if(context.is_compute)
-        {
-            return DoneResult(FormatComputeKernels(context, ASSISTANT_TOP_KERNEL_LIMIT),
-                              "Read compute kernel stats");
-        }
-        const AssistantFetchState summary_fetch =
-            TableFetch(AssistantFetchKind::kSummary, TableType::kSummaryKernelTable,
-                       ASSISTANT_DEFAULT_ROW_LIMIT);
-        if(context.data_provider->IsRequestPending(DataProvider::SUMMARY_REQUEST_ID))
-        {
-            return PendingResult(DataProvider::SUMMARY_REQUEST_ID, summary_fetch,
-                                 "Loading summary...", false);
-        }
-        const SummaryInfo::GPUMetrics& gpu =
-            context.data_provider->DataModel().GetSummary().GetSummaryData().gpu;
-        if(!gpu.top_kernels.empty() || gpu.kernel_exec_time_total > 0.0)
-        {
-            return DoneResult(FormatSystemSummary(context), "Read summary");
-        }
-        if(context.data_provider->FetchSummary())
-        {
-            return PendingResult(DataProvider::SUMMARY_REQUEST_ID, summary_fetch,
-                                 "Loading summary...", true);
-        }
-        if(context.data_provider->IsRequestPending(DataProvider::SUMMARY_REQUEST_ID))
-        {
-            return PendingResult(DataProvider::SUMMARY_REQUEST_ID, summary_fetch,
-                                 "Loading summary...", false);
-        }
-        return DoneResult(FormatSystemSummary(context), "Read summary");
-    }
-
-    if(tool_name == "list_tracks")
-    {
-        if(context.is_compute)
-        {
-            return DoneResult("Compute traces have no timeline tracks. Use kernel_metrics.",
-                              "No timeline tracks");
-        }
-        const std::vector<const TrackInfo*> tracks =
-            context.data_provider->DataModel().GetTimeline().GetTrackList();
-        std::ostringstream out;
-        size_t             shown = 0;
-        for(const TrackInfo* track : tracks)
-        {
-            if(track == nullptr)
-            {
-                continue;
-            }
-            if(track->topology.type != TrackInfo::Queue &&
-               track->topology.type != TrackInfo::Stream &&
-               track->topology.type != TrackInfo::InstrumentedThread &&
-               track->topology.type != TrackInfo::SampledThread &&
-               track->topology.type != TrackInfo::Counter)
-            {
-                continue;
-            }
-            out << "track_id=" << track->id << " type=" << TrackTypeName(track->topology.type)
-                << " reads_with="
-                << (track->track_type == kRPVControllerTrackTypeEvents ? "track_events"
-                                                                      : "track_samples")
-                << " name=\""
-                << context.data_provider->DataModel().BuildTrackName(track->id) << "\"";
-            out << " ops=";
-            bool first = true;
-            AppendOpName(out, first, kRocProfVisDmOperationDispatch, *track);
-            AppendOpName(out, first, kRocProfVisDmOperationMemoryCopy, *track);
-            AppendOpName(out, first, kRocProfVisDmOperationMemoryAllocate, *track);
-            AppendOpName(out, first, kRocProfVisDmOperationLaunch, *track);
-            AppendOpName(out, first, kRocProfVisDmOperationLaunchSample, *track);
-            if(first)
-            {
-                out << "-";
-            }
-            out << "\n";
-            ++shown;
-            if(shown >= ASSISTANT_MAX_LIST_TRACKS)
-            {
-                out << "... additional tracks omitted\n";
-                break;
-            }
-        }
-        if(shown == 0)
-        {
-            return DoneResult("No queue/thread tracks found.", "No tracks");
-        }
-        return DoneResult(TrimResult(out.str()), "Listed tracks");
-    }
-
-    if(tool_name == "goto")
-    {
-        const double start_ns =
-            JsonUtils::GetDouble(args, "start_ns",
-                                 JsonUtils::GetDouble(args, "start", -1.0));
-        const double end_ns =
-            JsonUtils::GetDouble(args, "end_ns", JsonUtils::GetDouble(args, "end", -1.0));
-        const uint64_t track_id =
-            JsonU64(args, "track_id", TimelineSelection::INVALID_SELECTION_ID);
-        const uint64_t event_uuid =
-            JsonU64(args, "event_uuid",
-                    JsonU64(args, "event_id", TimelineSelection::INVALID_SELECTION_ID));
-        const std::string kernel_name = JsonUtils::GetString(args, "kernel_name", "");
-
-        if(context.is_compute)
-        {
-            if(context.compute_selection == nullptr)
-            {
-                return DoneResult("Compute selection is not available.", "goto failed");
-            }
-            ComputeDataModel& model = context.data_provider->ComputeModel();
-            uint32_t workload_id    = context.compute_selection->GetSelectedWorkload();
-            const WorkloadInfo* workload = model.GetWorkload(workload_id);
-            if(workload == nullptr && !model.GetWorkloadList().empty())
-            {
-                workload = model.GetWorkloadList().front();
-            }
-            if(workload == nullptr)
-            {
-                return DoneResult("No compute workload is loaded.", "goto failed");
-            }
-            const uint32_t kernel_id = static_cast<uint32_t>(
-                JsonU64(args, "kernel_id", ComputeSelection::INVALID_SELECTION_ID));
-            const KernelInfo* kernel = FindComputeKernel(*workload, kernel_name, kernel_id);
-            if(kernel == nullptr)
-            {
-                return DoneResult("Could not find that kernel to select.", "goto failed");
-            }
-            Actions(context).SelectKernel(kernel->id);
-            return DoneResult(std::string("Selected compute kernel ") + kernel->name,
-                              "Selected kernel");
-        }
-
-        if(start_ns < 0.0 || end_ns < 0.0 || end_ns <= start_ns)
-        {
-            return DoneResult(
-                "goto needs start_ns and end_ns (end > start) from kernel_instances "
-                "or the selected range.",
-                "goto failed");
-        }
-        if(context.timeline_selection == nullptr)
-        {
-            return DoneResult("Timeline selection is not available.", "goto failed");
-        }
-
-        OptiqActions actions = Actions(context);
-        actions.SelectRange(start_ns, end_ns);
-        const bool zoom = JsonUtils::GetBool(args, "zoom", false);
-        if(zoom)
-        {
-            actions.ZoomToRange(start_ns, end_ns);
-        }
-
-        // The model may name one event via track_id/event_uuid or several via
-        // events[]; both shapes collapse to this list.
-        std::vector<std::pair<uint64_t, uint64_t>> targets;
-        if(event_uuid != TimelineSelection::INVALID_SELECTION_ID)
-        {
-            targets.emplace_back(track_id, event_uuid);
-        }
-        jt::Json& mutable_args = const_cast<jt::Json&>(args);
-        if(mutable_args.contains("events") && mutable_args["events"].isArray())
-        {
-            for(jt::Json& entry : mutable_args["events"].getArray())
-            {
-                if(!entry.isObject() || targets.size() >= ASSISTANT_MAX_HIGHLIGHTS)
-                {
-                    continue;
-                }
-                const uint64_t entry_uuid =
-                    JsonU64(entry, "event_uuid",
-                            JsonU64(entry, "__uuid",
-                                    TimelineSelection::INVALID_SELECTION_ID));
-                if(entry_uuid != TimelineSelection::INVALID_SELECTION_ID)
-                {
-                    targets.emplace_back(JsonU64(entry, "track_id", track_id), entry_uuid);
-                }
-            }
-        }
-
-        if(targets.empty())
-        {
-            actions.ShowRange(start_ns, end_ns);
-        }
-        else
-        {
-            actions.NavigateToEvent(targets.front().first, targets.front().second,
-                                    start_ns, end_ns - start_ns);
-        }
-
-        // Clicking the first target is what makes the trace view load its
-        // details, flow arrows, and call stack. NavigateToEvent only highlights,
-        // so without this the arrows never appear.
-        size_t highlighted = 0;
-        if(!targets.empty())
-        {
-            actions.ClearHighlights();
-            for(const std::pair<uint64_t, uint64_t>& target : targets)
-            {
-                if(!actions.HighlightEvent(target.first, target.second))
-                {
-                    continue;
-                }
-                if(highlighted == 0)
-                {
-                    actions.ClickEvent(target.first, target.second);
-                    actions.ScrollToTrack(target.first);
-                }
-                ++highlighted;
-            }
-        }
-
-        SettingsManager& settings    = SettingsManager::GetInstance();
-        const TimeFormat time_format = settings.GetUserSettings().unit_settings.time_format;
-        std::ostringstream out;
-        out << "Moved the timeline to "
-            << nanosecond_to_formatted_str(start_ns, time_format, true) << " .. "
-            << nanosecond_to_formatted_str(end_ns, time_format, true);
-        if(zoom)
-        {
-            out << ", zoomed in";
-        }
-        if(highlighted > 0)
-        {
-            out << " and highlighted " << highlighted
-                << " event(s) there, with flow arrows expanded on the first";
-        }
-        out << ".";
-        return DoneResult(out.str(), "Moved the view");
-    }
-
-    if(tool_name == "top_events")
-    {
-        if(context.is_compute)
-        {
-            return DoneResult(
-                "top_events is for system traces. Use get_summary / kernel_metrics.",
-                "Wrong trace kind");
-        }
-        const std::string category = JsonUtils::GetString(args, "category", "dispatch");
-        const TopEventsSpec* spec  = FindTopEventsSpec(category);
-        if(spec == nullptr)
-        {
-            return DoneResult(
-                "Unknown category. Use dispatch, memory_copy, memory_alloc, "
-                "instrumented, or sampled.",
-                "Bad category");
-        }
-        const size_t limit =
-            ClampRowLimit(JsonUtils::GetInt(args, "limit",
-                                            static_cast<int32_t>(ASSISTANT_DEFAULT_ROW_LIMIT)));
-
-        std::string track_error;
-        const std::vector<uint64_t> tracks =
-            TracksFromArgs(context, args, spec->op, true, track_error);
-        if(!track_error.empty())
-        {
-            return DoneResult(track_error, "Bad track_ids");
-        }
-        if(tracks.empty())
-        {
-            return DoneResult(
-                std::string("No tracks contain ") + spec->key +
-                    " events. Call list_tracks or try another category.",
-                "No matching tracks");
-        }
-
-        std::string       query_error;
-        const std::string where    = BuildWhereClause(args, query_error);
-        const std::string group_by = GroupByFromArgs(args, query_error);
-        if(!query_error.empty())
-        {
-            return DoneResult(query_error, "Bad query argument");
-        }
-
-        double start_ns = 0.0;
-        double end_ns   = 0.0;
-        TimeRangeFromArgs(context, args, start_ns, end_ns);
-
-        const AssistantFetchState fetch =
-            TableFetch(AssistantFetchKind::kTopEvents, spec->table_type, limit);
-        if(context.data_provider->IsRequestPending(spec->request_id))
-        {
-            return PendingResult(spec->request_id, fetch,
-                                 AssistantToolStatusLabel(tool_name), false);
-        }
-
-        const TablesModel& tables =
-            context.data_provider->DataModel().GetAnalysis().GetTables();
-        const bool queued = context.data_provider->FetchTable(TrackTableRequestParams(
-            spec->controller_type, tracks, start_ns, end_ns, where.c_str(), "",
-            group_by.c_str(), "", JsonU64(args, "offset", 0),
-            static_cast<uint64_t>(limit),
-            ResolveSortColumn(tables, spec->table_type,
-                              JsonUtils::GetString(args, "sort_by", ""),
-                              ASSISTANT_DURATION_COLUMN),
-            SortOrderFromArgs(args, kRPVControllerSortOrderDescending)));
-        if(!queued && !context.data_provider->IsRequestPending(spec->request_id))
-        {
-            return DoneResult("Could not queue the top_events fetch.", "Fetch failed");
-        }
-        return PendingResult(spec->request_id, fetch,
-                             AssistantToolStatusLabel(tool_name), queued);
-    }
-
-    if(tool_name == "kernel_instances")
-    {
-        if(context.is_compute)
-        {
-            return DoneResult(
-                "kernel_instances is for system traces. Use kernel_metrics on compute traces.",
-                "Wrong trace kind");
-        }
-        const std::string kernel_name = JsonUtils::GetString(args, "kernel_name", "");
-        if(kernel_name.empty())
-        {
-            return DoneResult("kernel_instances needs kernel_name.", "Missing kernel_name");
-        }
-        const size_t limit =
-            ClampRowLimit(JsonUtils::GetInt(args, "limit",
-                                            static_cast<int32_t>(ASSISTANT_DEFAULT_ROW_LIMIT)));
-
-        std::string       query_error;
-        const std::string where = BuildWhereClause(args, query_error);
-        if(!query_error.empty())
-        {
-            return DoneResult(query_error, "Bad query argument");
-        }
-
-        double start_ns = 0.0;
-        double end_ns   = 0.0;
-        TimeRangeFromArgs(context, args, start_ns, end_ns);
-
-        const uint64_t request_id = DataProvider::SUMMARY_KERNEL_INSTANCE_TABLE_REQUEST_ID;
-        const AssistantFetchState fetch =
-            TableFetch(AssistantFetchKind::kKernelInstances, TableType::kSummaryKernelTable,
-                       limit);
-        if(context.data_provider->IsRequestPending(request_id))
-        {
-            return PendingResult(request_id, fetch, AssistantToolStatusLabel(tool_name),
-                                 false);
-        }
-
-        const TablesModel& tables = context.data_provider->DataModel().GetTables();
-        const bool queued = context.data_provider->FetchTable(EventSearchRequestParams(
-            kRPVControllerTableTypeSummaryKernelInstances,
-            { kRocProfVisDmOperationDispatch }, start_ns, end_ns, where.c_str(), false,
-            false, false, { kernel_name }, JsonU64(args, "offset", 0),
-            static_cast<uint64_t>(limit),
-            ResolveSortColumn(tables, TableType::kSummaryKernelTable,
-                              JsonUtils::GetString(args, "sort_by", ""), 0),
-            SortOrderFromArgs(args, kRPVControllerSortOrderAscending)));
-        if(!queued && !context.data_provider->IsRequestPending(request_id))
-        {
-            return DoneResult("Could not queue the kernel_instances fetch.", "Fetch failed");
-        }
-        return PendingResult(request_id, fetch, AssistantToolStatusLabel(tool_name),
-                             queued);
-    }
-
-    if(tool_name == "kernel_metrics")
-    {
-        const std::string kernel_name = JsonUtils::GetString(args, "kernel_name", "");
-        const std::string metric_name = JsonUtils::GetString(args, "metric_name", "");
-        if(!context.is_compute)
-        {
-            std::ostringstream out;
-            const std::vector<SummaryInfo::KernelMetrics>& kernels =
-                context.data_provider->DataModel().GetSummary().GetSummaryData().gpu.top_kernels;
-            bool found = false;
-            for(const SummaryInfo::KernelMetrics& kernel : kernels)
-            {
-                if(kernel_name.empty() || kernel.name == kernel_name ||
-                   ToLowerCopy(kernel.name).find(ToLowerCopy(kernel_name)) !=
-                       std::string::npos)
-                {
-                    out << "name=" << kernel.name << " calls=" << kernel.invocations
-                        << " sum_ns=" << kernel.exec_time_sum
-                        << " min_ns=" << kernel.exec_time_min
-                        << " max_ns=" << kernel.exec_time_max
-                        << " pct=" << kernel.exec_time_pct << "\n";
-                    found = true;
-                    if(!kernel_name.empty())
-                    {
-                        break;
-                    }
-                }
-            }
-            if(!found)
-            {
-                out << "Kernel not in the current summary. Call get_summary first.\n";
-            }
-            return DoneResult(out.str(), "Read kernel summary stats");
-        }
-
-        ComputeDataModel& model = context.data_provider->ComputeModel();
-        uint32_t workload_id    = ComputeSelection::INVALID_SELECTION_ID;
-        if(context.compute_selection != nullptr)
-        {
-            workload_id = context.compute_selection->GetSelectedWorkload();
-        }
-        const WorkloadInfo* workload = model.GetWorkload(workload_id);
-        if(workload == nullptr && !model.GetWorkloadList().empty())
-        {
-            workload = model.GetWorkloadList().front();
-        }
-        if(workload == nullptr)
-        {
-            return DoneResult("No compute workload is loaded.", "No workload");
-        }
-
-        uint32_t selected_kernel = ComputeSelection::INVALID_SELECTION_ID;
-        if(context.compute_selection != nullptr)
-        {
-            selected_kernel = context.compute_selection->GetSelectedKernel();
-        }
-        const uint32_t kernel_id_arg = static_cast<uint32_t>(
-            JsonU64(args, "kernel_id", selected_kernel));
-        const KernelInfo* kernel =
-            FindComputeKernel(*workload, kernel_name, kernel_id_arg);
-        if(kernel == nullptr && selected_kernel != ComputeSelection::INVALID_SELECTION_ID)
-        {
-            kernel = FindComputeKernel(*workload, "", selected_kernel);
-        }
-        if(kernel == nullptr)
-        {
-            return DoneResult("Could not find that kernel. Call get_summary.",
-                              "Unknown kernel");
-        }
-
-        std::ostringstream out;
-        out << "kernel_id=" << kernel->id << " name=" << kernel->name << "\n";
-        out << "invocation_count=" << kernel->dispatch_metrics[KernelInfo::InvocationCount]
-            << "\n";
-        out << "duration_total_ns=" << kernel->dispatch_metrics[KernelInfo::DurationTotal]
-            << "\n";
-        out << "duration_min_ns=" << kernel->dispatch_metrics[KernelInfo::DurationMin]
-            << "\n";
-        out << "duration_max_ns=" << kernel->dispatch_metrics[KernelInfo::DurationMax]
-            << "\n";
-        out << "duration_mean_ns=" << kernel->dispatch_metrics[KernelInfo::DurationMean]
-            << "\n";
-        out << "duration_median_ns="
-            << kernel->dispatch_metrics[KernelInfo::DurationMedian] << "\n";
-
-        if(metric_name.empty())
-        {
-            out << "available_metrics:\n";
-            size_t listed = 0;
-            for(const AvailableMetrics::Entry& entry : workload->available_metrics.list)
-            {
-                out << "  " << entry.name;
-                if(!entry.unit.empty())
-                {
-                    out << " (" << entry.unit << ")";
-                }
-                out << "\n";
-                ++listed;
-                if(listed >= ASSISTANT_MAX_METRIC_NAMES)
-                {
-                    out << "  ... more metrics omitted; pass metric_name to fetch a subset\n";
-                    break;
-                }
-            }
-            return DoneResult(TrimResult(out.str()), "Read kernel dispatch stats");
-        }
-
-        const std::string needle = ToLowerCopy(metric_name);
-        std::vector<MetricsRequestParams::MetricID> metric_ids;
-        for(const AvailableMetrics::Entry& entry : workload->available_metrics.list)
-        {
-            if(ToLowerCopy(entry.name).find(needle) == std::string::npos)
-            {
-                continue;
-            }
-            MetricsRequestParams::MetricID id;
-            id.category_id = entry.category_id;
-            id.table_id    = entry.table_id;
-            id.entry_id    = entry.id;
-            metric_ids.push_back(id);
-            if(metric_ids.size() >= ASSISTANT_MAX_METRIC_FETCH)
-            {
-                break;
-            }
-        }
-        if(metric_ids.empty())
-        {
-            out << "No available metric name contains \"" << metric_name << "\".\n";
-            return DoneResult(TrimResult(out.str()), "Metric not found");
-        }
-
-        model.ClearKernelMetricValues(context.metrics_client_id);
-        const uint64_t request_id = RequestIdBuilder::MakeClientRequestId(
-            RequestType::kFetchMetrics, context.metrics_client_id);
-        AssistantFetchState metrics_fetch =
-            TableFetch(AssistantFetchKind::kMetrics, TableType::kSummaryKernelTable,
-                       ASSISTANT_DEFAULT_ROW_LIMIT);
-        metrics_fetch.kernel_id = kernel->id;
-        if(context.data_provider->IsRequestPending(request_id))
-        {
-            return PendingResult(request_id, metrics_fetch,
-                                 AssistantToolStatusLabel(tool_name), false);
-        }
-        const bool queued = context.data_provider->FetchMetrics(MetricsRequestParams(
-            workload->id, { kernel->id }, metric_ids, context.metrics_client_id));
-        if(!queued && !context.data_provider->IsRequestPending(request_id))
-        {
-            out << "Could not queue the metric fetch.\n";
-            return DoneResult(TrimResult(out.str()), "Fetch failed");
-        }
-        AssistantToolStartResult pending = PendingResult(
-            request_id, metrics_fetch, AssistantToolStatusLabel(tool_name), queued);
-        pending.content = out.str();
-        return pending;
-    }
-
-    if(tool_name == "track_events" || tool_name == "track_samples")
-    {
-        if(context.is_compute)
-        {
-            return DoneResult(
-                "Compute traces have no timeline tracks. Use kernel_metrics.",
-                "Wrong trace kind");
-        }
-
-        const bool      events   = tool_name == "track_events";
-        const TableType type     = events ? TableType::kEventTable : TableType::kSampleTable;
-        const uint64_t request_id = events ? DataProvider::EVENT_TABLE_REQUEST_ID
-                                           : DataProvider::SAMPLE_TABLE_REQUEST_ID;
-        const rocprofvis_controller_track_type_t wanted =
-            events ? kRPVControllerTrackTypeEvents : kRPVControllerTrackTypeSamples;
-
-        std::string           track_error;
-        std::vector<uint64_t> tracks = TracksFromArgs(
-            context, args, kRocProfVisDmOperationDispatch, false, track_error);
-        if(!track_error.empty())
-        {
-            return DoneResult(track_error, "Bad track_ids");
-        }
-        if(tracks.empty())
-        {
-            tracks = TracksOfType(context, wanted);
-        }
-        else
-        {
-            tracks = KeepTracksOfType(context, tracks, wanted);
-            if(tracks.empty())
-            {
-                return DoneResult(
-                    std::string("Those track_ids are not ") +
-                        (events ? "event tracks. Use track_samples for counter tracks."
-                                : "counter tracks. Use track_events for queue, stream, "
-                                  "and thread tracks."),
-                    "Wrong track kind");
-            }
-        }
-        if(tracks.empty())
-        {
-            return DoneResult(std::string("No ") +
-                                  (events ? "event" : "counter") +
-                                  " tracks found. Call list_tracks.",
-                              "No matching tracks");
-        }
-
-        std::string       query_error;
-        const std::string where    = BuildWhereClause(args, query_error);
-        const std::string group_by = GroupByFromArgs(args, query_error);
-        if(!query_error.empty())
-        {
-            return DoneResult(query_error, "Bad query argument");
-        }
-
-        const size_t limit =
-            ClampRowLimit(JsonUtils::GetInt(args, "limit",
-                                            static_cast<int32_t>(ASSISTANT_DEFAULT_ROW_LIMIT)));
-        double start_ns = 0.0;
-        double end_ns   = 0.0;
-        TimeRangeFromArgs(context, args, start_ns, end_ns);
-
-        const AssistantFetchState fetch =
-            TableFetch(AssistantFetchKind::kDataTable, type, limit);
-        if(context.data_provider->IsRequestPending(request_id))
-        {
-            return PendingResult(request_id, fetch, AssistantToolStatusLabel(tool_name),
-                                 false);
-        }
-
-        const TablesModel& tables = context.data_provider->DataModel().GetTables();
-        const bool queued = context.data_provider->FetchTable(TrackTableRequestParams(
-            events ? kRPVControllerTableTypeEvents : kRPVControllerTableTypeSamples,
-            tracks, start_ns, end_ns, where.c_str(), "", group_by.c_str(), "",
-            JsonU64(args, "offset", 0), static_cast<uint64_t>(limit),
-            ResolveSortColumn(tables, type, JsonUtils::GetString(args, "sort_by", ""), 0),
-            SortOrderFromArgs(args, kRPVControllerSortOrderAscending)));
-        if(!queued && !context.data_provider->IsRequestPending(request_id))
-        {
-            return DoneResult("Could not queue the " + tool_name + " fetch.",
-                              "Fetch failed");
-        }
-        return PendingResult(request_id, fetch, AssistantToolStatusLabel(tool_name),
-                             queued);
-    }
-
-    if(tool_name == "event_details")
-    {
-        if(context.is_compute)
-        {
-            return DoneResult("event_details is for system traces.", "Wrong trace kind");
-        }
-        const uint64_t event_uuid =
-            JsonU64(args, "event_uuid", JsonU64(args, "event_id", 0));
-        if(event_uuid == 0)
-        {
-            return DoneResult("event_details needs event_uuid from the __uuid column.",
-                              "Missing event_uuid");
-        }
-
-        const std::vector<uint64_t> request_ids = {
-            DataProvider::EVENT_EXTENDED_DATA_REQUEST_ID,
-            DataProvider::EVENT_FLOW_DATA_REQUEST_ID,
-            DataProvider::EVENT_CALL_STACK_DATA_REQUEST_ID
-        };
-        // These request ids are shared with the event panel, and re-issuing one
-        // that is already in flight would drop its future on the floor.
-        for(uint64_t request_id : request_ids)
-        {
-            if(context.data_provider->IsRequestPending(request_id))
-            {
-                return DoneResult("Event details are already loading. Try again.",
-                                  "Busy");
-            }
-        }
-
-        const uint64_t track_id = JsonU64(args, "track_id", INVALID_UINT64_INDEX);
-        // FetchEvent seeds name/duration from loaded track data and chains the
-        // extended-data fetch; without a track id only the fetches below run.
-        bool extended = false;
-        if(track_id != INVALID_UINT64_INDEX)
-        {
-            extended = context.data_provider->FetchEvent(track_id, event_uuid);
-        }
-        const bool flow  = context.data_provider->FetchEventFlowDetails(event_uuid);
-        const bool stack = context.data_provider->FetchEventCallStackData(event_uuid);
-        if(!extended && !flow && !stack)
-        {
-            return DoneResult("Could not load details for that event.", "Fetch failed");
-        }
-
-        AssistantFetchState fetch;
-        fetch.kind     = AssistantFetchKind::kEventDetails;
-        fetch.event_id = event_uuid;
-        fetch.track_id = track_id;
-        return PendingResult(request_ids, fetch, AssistantToolStatusLabel(tool_name),
-                             true);
-    }
-
-    if(tool_name == "track_statistics")
-    {
-        if(context.is_compute)
-        {
-            return DoneResult("track_statistics is for system traces.",
-                              "Wrong trace kind");
-        }
-        const uint64_t track_id = JsonU64(args, "track_id", INVALID_UINT64_INDEX);
-        const TrackInfo* track =
-            context.data_provider->DataModel().GetTimeline().GetTrack(track_id);
-        if(track == nullptr)
-        {
-            return DoneResult("Unknown track_id. Call list_tracks first.",
-                              "Unknown track");
-        }
-        if(track->topology.type != TrackInfo::Queue &&
-           track->topology.type != TrackInfo::Counter)
-        {
-            return DoneResult(
-                "track_statistics only works on Queue and Counter tracks. Use "
-                "track_events for thread and stream tracks.",
-                "Wrong track type");
-        }
-
-        const AnalysisTrackStatistics* stats =
-            context.data_provider->DataModel().GetAnalysis().RegisterTrack(*track);
-        if(stats == nullptr)
-        {
-            return DoneResult("Statistics are not available for that track.",
-                              "No statistics");
-        }
-
-        double start_ns = 0.0;
-        double end_ns   = 0.0;
-        TimeRangeFromArgs(context, args, start_ns, end_ns);
-
-        AssistantFetchState fetch;
-        fetch.kind     = AssistantFetchKind::kTrackStatistics;
-        fetch.track_id = track_id;
-
-        const uint64_t request_id = RequestIdBuilder::MakeTrackDataRequestId(
-            static_cast<uint32_t>(track_id), 0, 0,
-            RequestType::kFetchAnalysisTrackStatistics);
-        if(context.data_provider->IsRequestPending(request_id))
-        {
-            return PendingResult(request_id, fetch, AssistantToolStatusLabel(tool_name),
-                                 false);
-        }
-
-        const bool queued = context.data_provider->FetchAnalysisTrackStatistics(
-            AnalysisTrackStatisticsRequestParams(track_id, start_ns, end_ns));
-        stats->state = queued ? AnalysisTrackStatistics::kRequested
-                              : AnalysisTrackStatistics::kPending;
-        if(!queued)
-        {
-            return DoneResult("Could not queue the track statistics fetch.",
-                              "Fetch failed");
-        }
-        return PendingResult(request_id, fetch, AssistantToolStatusLabel(tool_name), true);
-    }
-
-    if(tool_name == "search_events")
-    {
-        if(context.is_compute)
-        {
-            return DoneResult(
-                "search_events is for system traces. Use kernel_metrics on compute "
-                "traces.",
-                "Wrong trace kind");
-        }
-
-        std::vector<std::string> terms = JsonUtils::GetStringArray(args, "terms");
-        const std::string        single = JsonUtils::GetString(args, "query", "");
-        if(terms.empty() && !single.empty())
-        {
-            terms.push_back(single);
-        }
-        if(terms.empty())
-        {
-            return DoneResult(
-                "search_events needs terms, e.g. {\"terms\":[\"gemm\"]}.",
-                "Missing terms");
-        }
-        if(terms.size() > ASSISTANT_MAX_SEARCH_TERMS)
-        {
-            terms.resize(ASSISTANT_MAX_SEARCH_TERMS);
-        }
-
-        std::vector<rocprofvis_dm_event_operation_t> ops;
-        for(const std::string& category : JsonUtils::GetStringArray(args, "categories"))
-        {
-            const TopEventsSpec* spec = FindTopEventsSpec(category);
-            if(spec == nullptr)
-            {
-                return DoneResult("Unknown category \"" + category +
-                                      "\". Use dispatch, memory_copy, memory_alloc, "
-                                      "instrumented, or sampled.",
-                                  "Bad category");
-            }
-            ops.push_back(spec->op);
-        }
-        if(ops.empty())
-        {
-            ops = { kRocProfVisDmOperationLaunch, kRocProfVisDmOperationDispatch,
-                    kRocProfVisDmOperationMemoryCopy,
-                    kRocProfVisDmOperationMemoryAllocate,
-                    kRocProfVisDmOperationLaunchSample };
-        }
-
-        const bool contains =
-            ToLowerCopy(JsonUtils::GetString(args, "match", "contains")) != "equals";
-        const bool any_term =
-            ToLowerCopy(JsonUtils::GetString(args, "combine", "all")) == "any";
-        const bool include_category =
-            JsonUtils::GetBool(args, "include_category", false);
-
-        std::string       query_error;
-        const std::string where = BuildWhereClause(args, query_error);
-        if(!query_error.empty())
-        {
-            return DoneResult(query_error, "Bad query argument");
-        }
-
-        const size_t limit =
-            ClampRowLimit(JsonUtils::GetInt(args, "limit",
-                                            static_cast<int32_t>(ASSISTANT_DEFAULT_ROW_LIMIT)));
-        double start_ns = 0.0;
-        double end_ns   = 0.0;
-        TimeRangeFromArgs(context, args, start_ns, end_ns, false);
-
-        const uint64_t            request_id = DataProvider::EVENT_SEARCH_REQUEST_ID;
-        const AssistantFetchState fetch = TableFetch(
-            AssistantFetchKind::kDataTable, TableType::kEventSearchTable, limit);
-        if(context.data_provider->IsRequestPending(request_id))
-        {
-            return PendingResult(request_id, fetch, AssistantToolStatusLabel(tool_name),
-                                 false);
-        }
-
-        const TablesModel& tables = context.data_provider->DataModel().GetTables();
-        const bool queued = context.data_provider->FetchTable(EventSearchRequestParams(
-            kRPVControllerTableTypeSearchResults, ops, start_ns, end_ns, where.c_str(),
-            contains, include_category, any_term, terms, JsonU64(args, "offset", 0),
-            static_cast<uint64_t>(limit),
-            ResolveSortColumn(tables, TableType::kEventSearchTable,
-                              JsonUtils::GetString(args, "sort_by", ""), 1),
-            SortOrderFromArgs(args, kRPVControllerSortOrderAscending)));
-        if(!queued && !context.data_provider->IsRequestPending(request_id))
-        {
-            return DoneResult("Could not queue the search.", "Fetch failed");
-        }
-        return PendingResult(request_id, fetch, AssistantToolStatusLabel(tool_name),
-                             queued);
-    }
-
-    return DoneResult("Unknown tool. Use one of: " + AssistantToolNameList() + ".",
+    return DoneResult("Unknown tool \"" + tool_name + "\". Available tools: " +
+                          AssistantToolNameList() + ".",
                       "Unknown tool");
 }
