+++ src/view/src/rocprofvis_settings_manager.cpp
@@ -537,6 +537,9 @@ SettingsManager::LoadSettingsJson()
         DeserializeProfilerSettings(result.second);
         DeserializeAssistantSettings(result.second);
         DeserializeAppWindowSettings(result.second);
+#ifdef ROCPROFVIS_ENABLE_AGENTIC_PROFILING
+        MigrateLegacyAssistantToken();
+#endif
     }
     else
     {
@@ -1162,8 +1165,8 @@ SettingsManager::DeserializeAppWindowSettings(jt::Json& json)
 #ifdef ROCPROFVIS_ENABLE_AGENTIC_PROFILING
 namespace
 {
-// Credential-store entry for one provider. Keys written before routes existed
-// live under the bare prefix, which GetAssistantToken still falls back to.
+// Credential-store entry for one endpoint. Every endpoint gets its own entry,
+// so a key configured for one host can never be sent to another.
 std::string
 AssistantTokenKey(const std::string& provider_name)
 {
@@ -1175,6 +1178,48 @@ AssistantTokenKey(const std::string& provider_name)
 }
 }  // namespace
 
+// Moves a key written before endpoints were named onto the first endpoint, so
+// an upgrade does not make the user re-enter it.
+//
+// This runs once at load and then deletes the unnamed entry. Reading it as a
+// live fallback instead would mean any endpoint without a key of its own
+// inherited this one, which is how a key for one provider would end up being
+// posted to a different provider's host.
+void
+SettingsManager::MigrateLegacyAssistantToken()
+{
+    if(!SecretStore::IsAvailable() || m_usersettings.assistant.providers.empty())
+    {
+        return;
+    }
+
+    std::string legacy_token;
+    if(!SecretStore::Get(ASSISTANT_TOKEN_SECRET_KEY, legacy_token) ||
+       legacy_token.empty())
+    {
+        return;
+    }
+
+    // The single endpoint that existed before the list did is the first one.
+    const std::string key =
+        AssistantTokenKey(m_usersettings.assistant.providers.front().name);
+    if(key == ASSISTANT_TOKEN_SECRET_KEY)
+    {
+        return;
+    }
+
+    std::string existing_token;
+    const bool  already_migrated =
+        SecretStore::Get(key, existing_token) && !existing_token.empty();
+    if(!already_migrated && !SecretStore::Set(key, legacy_token))
+    {
+        // Leave the old entry alone rather than destroying the only copy.
+        spdlog::warn("Could not migrate the saved assistant key; leaving it in place");
+        return;
+    }
+    SecretStore::Erase(ASSISTANT_TOKEN_SECRET_KEY);
+}
+
 // True when a key is already saved for this endpoint.
 bool
 SettingsManager::HasAssistantToken(const std::string& provider_name) const
@@ -1183,20 +1228,22 @@ SettingsManager::HasAssistantToken(const std::string& provider_name) const
     return GetAssistantToken(provider_name, unused);
 }
 
-// Reads the key from the credential store, this session's memory, or the
-// pre-provider entry, in that order.
+// Reads this endpoint's key from the credential store, or from this session's
+// memory when no store is available. Never falls through to another endpoint's
+// key: an endpoint with nothing saved has no key.
 bool
 SettingsManager::GetAssistantToken(const std::string& provider_name,
                                    std::string&       out_token) const
 {
     out_token.clear();
-    const std::string key = AssistantTokenKey(provider_name);
-    if(SecretStore::IsAvailable() && SecretStore::Get(key, out_token) &&
+    if(SecretStore::IsAvailable() &&
+       SecretStore::Get(AssistantTokenKey(provider_name), out_token) &&
        !out_token.empty())
     {
         return true;
     }
 
+    out_token.clear();
     const std::map<std::string, std::string>::const_iterator session =
         m_assistant_token_session.find(provider_name);
     if(session != m_assistant_token_session.end() && !session->second.empty())
@@ -1204,16 +1251,6 @@ SettingsManager::GetAssistantToken(const std::string& provider_name,
         out_token = session->second;
         return true;
     }
-
-    // Fall back to the key stored before providers were configurable, so an
-    // upgrade does not make the user re-enter it.
-    out_token.clear();
-    if(key != ASSISTANT_TOKEN_SECRET_KEY && SecretStore::IsAvailable() &&
-       SecretStore::Get(ASSISTANT_TOKEN_SECRET_KEY, out_token) && !out_token.empty())
-    {
-        return true;
-    }
-    out_token.clear();
     return false;
 }
 
@@ -1227,15 +1264,18 @@ SettingsManager::SetAssistantToken(const std::string& provider_name,
         return ClearAssistantToken(provider_name);
     }
 
-    m_assistant_token_session[provider_name] = token;
     if(SecretStore::IsAvailable())
     {
+        // The vault owns it now. A second copy in process memory would only
+        // widen the window where the key can be read back out.
+        m_assistant_token_session.erase(provider_name);
         return SecretStore::Set(AssistantTokenKey(provider_name), token);
     }
+    m_assistant_token_session[provider_name] = token;
     return true;
 }
 
-// Forgets the key everywhere it could be stored.
+// Forgets this endpoint's key. Other endpoints keep theirs.
 bool
 SettingsManager::ClearAssistantToken(const std::string& provider_name)
 {
@@ -1244,17 +1284,7 @@ SettingsManager::ClearAssistantToken(const std::string& provider_name)
     {
         return true;
     }
-
-    const std::string key     = AssistantTokenKey(provider_name);
-    const bool        erased  = SecretStore::Erase(key);
-    if(key == ASSISTANT_TOKEN_SECRET_KEY)
-    {
-        return erased;
-    }
-    // Also drop the pre-provider key, otherwise GetAssistantToken would fall
-    // back to it and the key would look like it survived being cleared.
-    SecretStore::Erase(ASSISTANT_TOKEN_SECRET_KEY);
-    return erased;
+    return SecretStore::Erase(AssistantTokenKey(provider_name));
 }
 #endif  // ROCPROFVIS_ENABLE_AGENTIC_PROFILING
