+++ src/view/src/rocprofvis_utils.h
@@ -304,5 +304,15 @@ is_remote_display_session();
 std::string
 strip_ansi_for_display(std::string const& text);
 
+/**
+ * @brief Returns a lowercased copy, for case-insensitive matching of names the
+ *        user or a tool argument supplied.
+ *
+ * Only ASCII is folded, which is what the callers need: they compare against
+ * panel names, tab labels, and column names that are all ASCII literals.
+ */
+std::string
+to_lower_copy(const std::string& value);
+
 }  // namespace View
 }  // namespace RocProfVis
